==> code/1_alb_data_manip.Rmd <==
---
title: "Step 1 Data Manip"
author: "Natasha Hardy"
date: "October 10, 2019"
output:
  html_document: default
  pdf_document: default
---

## About

This document contains code for manipulating trait information for 5 prey traits, and life history estimation, for albacore prey species from our global meta-analysis of albacore diet composition.

Relates to following sections of MS:

"2.2 Prey life stage estimation
See MS"

"2.3 Prey trait information

For each prey species and main life stage distinctions (i.e., larva, juvenile, adult), we collected trait information for four habitat use traits and one behavioural trait known to affect the likelihood of a pelagic predator encountering, attacking and capturing prey (Green et al., 2019). These were: (i) vertical and (ii) horizontal habitat association, (iii) presence of diel vertical migration, (iv) presence of seasonal migration and (v) aggregation behaviour (Supplementary Data, Table S4 & S5). For analyses, we include the phylogenetic Order verified using Open Tree of Life Data (OpenTree et al, 2020). Trait data were compiled for a broader database construction effort of traits that inform predator-prey interactions for albacore (Gleiber et al. in prep). We further describe how prey species’ trait values were collected in Appendix B (Supporting Information)."

## Workpsace

```{r setup, include=FALSE, echo=FALSE, message=FALSE}
knitr::opts_chunk$set(echo = TRUE, warning = FALSE)
```

```{r Load Workspace, message=FALSE}

#R documentation
library(pander)
# general
library(tidyverse)
library(readxl)
library(readr)
library(plyr)
library(dplyr)
library(devtools)
library(here)
"%notin%" = Negate('%in%')
here::here()

```

## Outputs

List of outputs created for analyses:

- Prey species' coarse trait information *data/2_output_data/prey_trait_select.csv*

- Albacore diet data *data/2_output_data/alb_global_diet.csv*

- Albacore diet sampling covariates *data/2_output_data/alb_covars.csv*

- And many more for supplementary information!

## Dataframes

### Diet raw data

Load diet data and merge with prey size trait information, and undertake necessary data selection and manipulation of variables.

Note: total dataset contained 2051 rows of individual observations of prey items from 32 studies.

```{r Albacore diet original data load & manipulation, tidy=TRUE}
#Load diet data spreadsheet
alb_global = as.data.frame(read_xlsx(here::here(
  "data/1_input_data/albacore_diet_global.xlsx"), 
  sheet = "albacore_diet_global")) #2051 rows and 77 vars
nrow(alb_global)
length(unique(alb_global$StudyID)) #32 studies
original = data.frame(unique(alb_global$StudyID))
```

Note: 3 studies investigated the diets of larval albacore tuna (Catalan et al. 2007, Uotani et al. 1982, and Young & Davis 1990) and one (Hart 1942) contained data that was also published again in a longer time series (Hart 1948). These four were thus excluded from further analyses of juvenile and adult tuna diets, with 1837 observations of prey items consumed from 28 studies.

```{r Albacore diet filter data by albacore life stage, tidy=TRUE}
alb_global = alb_global %>% 
  dplyr::filter(Include %in% c("Yes", "Maybe") & !pred_life == "larvae") 
nrow(alb_global)
length(unique(alb_global$StudyID)) #28 studies
check28 = data.frame(unique(alb_global$StudyID))

original_check28 = original %>%
  anti_join(check28)
unique(original_check28$unique.alb_global.StudyID.)
rm(original, original_check28)
```

Note: a further 2 studies did not report any prey items to species level and we did not include these in this meta-analysis. These were Logan et al (2013) and Teffer et al. (2015). Thus we retained 26 studies with 995 rows of prey species-level diet observations.

```{r Albacore diet data by prey species identified, tidy=TRUE}
alb_global = alb_global %>%
  dplyr::filter(TaxLev == "species") #995 rows and 77 vars
nrow(alb_global) 
length(unique(alb_global$StudyID)) #28 studies

check26 = data.frame(unique(alb_global$StudyID))

check26_check28 = check28 %>%
  anti_join(check26)
unique(check26_check28$unique.alb_global.StudyID.)
rm(check28, check26_check28, check26)
```

Data manipulations for global diet dataset for analyses:

```{r Albacore diet data manipulation, tidy=TRUE}

alb_global = alb_global %>%
  dplyr::select(StudyID, CiteAuth, CiteYear, OceanBasin, OceanBasinQ, LocatName,
                LocatLatitude:LocatLongitude, MonthRange, YearEnd, SampleMethod, SampleDepth,
                stomachs_used,
                pred_life:est_ref, stomachs_used, prey_class:prey_sp,
                prey_age_reported_1:prey_age_use, Include, FO, 
                pFOuse, N, pN, `Total Mass (g)`, pM, maxl_use, maxl_type) %>%
  #Add gape length information for albacore by study, to relate to potential prey length
  mutate(gape_lmax = pred_flmean*0.0823+1.758, 
         gape_height_limit = pred_flmean*0.0246+4.603,
         prey_age_reported_1=case_when(
                             #prey_age_reported_1 == "NA" ~ "none",
                             prey_age_reported_1 == "larvae" ~ "larva",
                             prey_age_reported_1 == "juvenile" ~ "juvenile",
                             prey_age_reported_1 == "adult" ~ "adult")) %>%
  arrange(StudyID)
  #1028 obsm 43 vars
  #calculates the gape length and height limits
  #Reference paper for this calculation -- Menard et al. (2006) for yellowfin tuna

#Correct data type
alb_global$maxl_use <- as.numeric(alb_global$maxl_use)

#The NA's in the reported age for prey column cause problems later and need 
#to be assigned a categorical value
alb_global$prey_age_reported_1 <- as.character(alb_global$prey_age_reported_1)
alb_global$prey_age_reported_1 <- replace_na(alb_global$prey_age_reported_1, "none")
#check this
unique(as.factor(alb_global$prey_age_reported_1))

#Change vector type for YearEnd variable
alb_global$YearEnd <- as.integer(alb_global$YearEnd)

#Deal with NA's in stomachs used column
alb_global$stomachs_used <- as.integer(alb_global$stomachs_used)
alb_global$stomachs_used <- as.character(alb_global$stomachs_used)
alb_global$stomachs_used <- replace_na(alb_global$stomachs_used, "not reported")
#Check this briefly
range(alb_global$stomachs_used)
length(unique(alb_global$prey_sp)) #308

#Save total df version
write.csv(alb_global, here::here("data/2_output_data/alb_global_diet.csv"), 
          row.names = FALSE) 
#1028 observations and 43 variables
```

### Covariates

Note that we lose two replicates here:

- McHugh 1952 (Mexico to USA) due to containing only higher level taxa classifications and not species-level data

- Williams 2015 (NZ_L_2010) due to data entry error or copy/paste error, data have been copied over.

There are 69 observations of albacore gut content data from 26 studies moving forward.

```{r Covariates df, tidy = TRUE}

alb_covars = alb_global %>%
  dplyr::select(StudyID, CiteYear, OceanBasin, OceanBasinQ, LocatName, LocatLatitude:pred_life_est) %>% 
  #, stomachs_used
  distinct(StudyID, CiteYear, OceanBasin, OceanBasinQ, LocatName, LocatLatitude, 
           LocatLongitude, YearEnd, MonthRange, stomachs_used, pred_life) %>% 
  #, stomachs_used
  mutate(lat_cat = if_else(abs(LocatLatitude) > 23.43662, "temperate", "tropical"),
         pred_life_adjust=case_when(pred_life == "juvenile" ~ "juvenile",
                             pred_life == "juvenile, adult" ~ "mixed",
                             pred_life == "none" ~ "mixed",
                             pred_life == "adult" ~ "adult"))

alb_covars$rep_ID = paste(alb_covars$StudyID, row.names(alb_covars), sep = "_")

length(unique(alb_covars$StudyID)) #26
length(unique(alb_covars$rep_ID)) #69
```

Save diet and covariate dfs:

```{r Save diet and covariate dfs}
#Merge these back to the diet df
alb_global_diet = alb_global %>%
  left_join(alb_covars)

write.csv(alb_covars, here::here("data/2_output_data/alb_covars.csv"), row.names = FALSE)
write.csv(alb_global_diet, here::here("data/2_output_data/alb_global_diet_total.csv"), 
          row.names = FALSE) 
#995 observations and 46 variables
```


### Table S3 content

```{r Covariates table S3, tidy = TRUE}

alb_table_S3 = alb_global %>%
  dplyr::select(StudyID, CiteYear, OceanBasin, OceanBasinQ, LocatName, LocatLatitude:est_ref) %>% 
  #, stomachs_used
  distinct(StudyID, CiteYear, OceanBasin, OceanBasinQ, LocatName, LocatLatitude, 
           LocatLongitude, YearEnd, MonthRange, SampleMethod, SampleDepth, stomachs_used, 
           pred_life, pred_life_est, pred_age, pred_age_est, pred_flmean, pred_flmean_se, pred_flmean_est, pred_flmin, pred_flmax, pred_flest, est_note, est_ref) %>% 
  #, stomachs_used
  mutate(#lat_cat = if_else(abs(LocatLatitude) > 23.43662, "temperate", "tropical"),
         pred_life_adjust=case_when(pred_life == "juvenile" ~ "juvenile",
                             pred_life == "juvenile, adult" ~ "mixed",
                             pred_life == "none" ~ "mixed",
                             pred_life == "adult" ~ "adult"),
         rep_ID = paste(alb_covars$StudyID, row.names(alb_covars), sep = "_"))

unique(alb_table_S3$StudyID) #26
unique(alb_table_S3$rep_ID) #69

write.csv(alb_table_S3, here::here("data/3_output_supp/table_S3.csv"), row.names = FALSE)
write.csv(alb_table_S3, here::here("outputs_figures/ms_figures/SI_results/table_S3.csv"), row.names = FALSE)

```

## Data checks & additional info

### Prey life stage reported

How many prey taxa were reported with an observation or estimation of life stage consumed?

```{r Reported prey lifestage information, tidy = TRUE}

alb_global_prey_life = alb_global %>%
  filter(prey_age_reported_1 != "none",) %>%
  dplyr::select(prey_class:prey_sp, prey_age_reported_1, pFOuse, pN, pM, maxl_use, maxl_type) %>%
  group_by(prey_sp, prey_age_reported_1, maxl_type) %>% #, life_stage , pFOuse, pN, pM
  dplyr::summarize(maxFO = max(pFOuse, na.rm=TRUE), 
                   maxN = max(pN, na.rm=TRUE), 
                   maxM = max(pM, na.rm=TRUE), 
                   maxL = max(maxl_use, na.rm=TRUE)) 

length(unique(alb_global_prey_life$prey_sp)) #72 species
#What percent of the total number of species?
length(unique(alb_global_prey_life$prey_sp))*100/308
#23.4%
#What number of species were identified to which life stage?
summary(as.factor(alb_global_prey_life$prey_age_reported_1))
#Note that 13 species had multiple life stage assignments (up to 3)

write.csv(alb_global_prey_life, here::here("data/3_output_supp/alb_table_E1_global_prey_life.csv"), row.names = FALSE)

```

### Prey length info reported

How many prey taxa were recorded with lengths of remains measured and thus able to calculate likely total lengths?

```{r Reported prey length information, tidy = TRUE}

alb_global_prey_lengths = alb_global %>%
  filter(maxl_use > 0) %>%
  dplyr::select(prey_class:prey_sp, pFOuse, pN, pM, maxl_use, maxl_type) %>%
  group_by(prey_sp, maxl_type) %>% #, life_stage , pFOuse, pN, pM
  dplyr::summarize(maxFO = max(pFOuse, na.rm=TRUE), 
                   maxN = max(pN, na.rm=TRUE), 
                   maxM = max(pM, na.rm=TRUE), 
                   maxL = max(maxl_use, na.rm=TRUE)) 

length(unique(alb_global_prey_lengths$prey_sp)) #43 species

write.csv(alb_global_prey_lengths, here::here("data/3_output_supp/alb_table_E2_global_prey_lengths.csv"), row.names = FALSE)
```

More precisely, how many prey taxa without reported life stage were recorded with lengths of remains measured and thus able to calculate likely total lengths? Note we exclude three Cephalopods here because their length measurements were for beaks and we did not at the time have any conversion calculations to convert those measurements to mantle or total lengths.

```{r check how much data this added}
check_prey_life = data.frame(unique(alb_global_prey_life$prey_sp)) %>%
  dplyr::rename(prey_sp = unique.alb_global_prey_life.prey_sp.)

check_length_info = alb_global_prey_lengths %>%
  anti_join(check_prey_life) %>%
  filter(!prey_sp %in% c("Onychoteuthis borealijaponica", "Onykia robusta", "Vampyroteuthis infernalis"))

length(unique(check_length_info$prey_sp))
length(unique(check_length_info$prey_sp))*100/308
#4.8

write.csv(check_length_info, here::here("data/3_output_supp/alb_table_E3_lengths_life_stage.csv"), row.names = FALSE)
```

How much data do we then have some near direct estimation capacity for life stage and how much is left to estimate?

```{r check how much data remained for further estimation}
#Species for which we have either reported life stage or measurement info
length(unique(alb_global_prey_life$prey_sp))+length(unique(check_length_info$prey_sp))
#=87
#Species without that
308-(length(unique(alb_global_prey_life$prey_sp))+length(unique(check_length_info$prey_sp)))
#=221
#Proportion of data with info
(length(unique(alb_global_prey_life$prey_sp))+length(unique(check_length_info$prey_sp)))*100/308
#28%
#Proportion of data without
(308-(length(unique(alb_global_prey_life$prey_sp))+length(unique(check_length_info$prey_sp))))*100/308
#72%
```

### Prey prevalence data reported

```{r Reported diet metrics summary v1, tidy = TRUE}

alb_global_metrics = alb_global %>%
  #unique(c(alb_global_prob$prey_sp, alb_global_prob$prey_select))
  dplyr::select(prey_class:prey_sp, prey_age_reported_1, pFOuse, pN, pM, maxl_use, maxl_type) %>%
  group_by(prey_sp, prey_age_reported_1, maxl_type) %>% #, life_stage , pFOuse, pN, pM
  dplyr::summarize(maxFO = max(pFOuse, na.rm=TRUE), 
                   maxN = max(pN, na.rm=TRUE), 
                   maxM = max(pM, na.rm=TRUE), 
                   maxL = max(maxl_use, na.rm=TRUE)) 

write.csv(alb_global_metrics, here::here("data/3_output_supp/albtable_Eglobal_metrics.csv"), row.names = FALSE)

```

```{r Reported diet metrics summary v2, tidy = TRUE}

alb_global_metrics_basic = alb_global %>%
  #unique(c(alb_global_prob$prey_sp, alb_global_prob$prey_select))
  dplyr::select(prey_class:prey_sp, pFOuse, pN, pM) %>%
  group_by(prey_sp) %>% #, life_stage , pFOuse, pN, pM
  dplyr::summarize(maxFO = max(pFOuse, na.rm=TRUE), 
                   maxN = max(pN, na.rm=TRUE), 
                   maxM = max(pM, na.rm=TRUE)) 

write.csv(alb_global_metrics_basic, here::here("data/3_output_supp/alb_table_E5_global_metrics_basic.csv"), row.names = FALSE)

```


## Prey trait db

### Load and check trait information

Loading in data. Note that NH overhauled our prey life stage estimation protocol in December 2021 through early 2022, and we have thoroughly curated this step. These steps are to investigate how much information was gained from this curation step.

```{r Prey trait data load, message=FALSE, results='hide', tidy=TRUE}

#Load prey trait db
prey_trait <- as.data.frame(read_xlsx(here::here("data/1_input_data/prey_trait_db.xlsx"),
                                           sheet = "prey_trait_db"))

```

Note that there are 308 spp remaining in the global diet db, and there are 221 for which we need to resolve which trait information to use. Why the difference in species from the trait db? There were 8 prey taxa found in the larval albacore tuna diets and were thus previously removed.

```{r check why some species drop out, tidy=TRUE}
check_spp = prey_trait %>%
  anti_join(alb_global_metrics_basic, by="prey_sp")
```

Need to filter by prey species we already have life stage or length data to estimate life stage. This leaves 221 species for which we need to figure out life stage information.

```{r Tally species remaining to id traits, tidy=TRUE}
check_traits = prey_trait %>%
  filter(!prey_sp %in% check_spp$prey_sp) %>%
  filter(!prey_sp %in% check_prey_life$prey_sp) %>%
  filter(!prey_sp %in% check_length_info$prey_sp) %>%
  select(prey_class:trait_confidence_cat, vert_habitat, horz_habitat, season_migrant, diel_migrant)

length(unique(check_traits$prey_sp)) #221.
```

What meta-information did I annotate for species? There were 65 with annotations (and 157 without), likely hinting at a decision that needed to be made. But of these some did include the same trait information for adult and juvenile life stages, so we can't be sure from these data alone.

```{r Checking on trait metainformation, tidy=TRUE}
length(which(check_traits$life_note != "NA")) #191

check_traits_a = check_traits %>%
  slice(which(check_traits$life_note != "NA"))
length(unique(check_traits_a$prey_sp))

check_traits_b = check_traits %>%
  slice(-which(check_traits$life_note != "NA"))
length(unique(check_traits_b$prey_sp))
```

Which species have the same or different vertical habitat use traits?

```{r Check same vert traits, message=FALSE}
df_vert = check_traits %>%
  select(prey_sp, life_stage, vert_habitat) %>%
  pivot_wider(., names_from = life_stage, values_from = vert_habitat)

check_traits_vert = df_vert %>%
  slice(which(df_vert$adult == df_vert$juvenile))
#136 with same

check_traits_vert_not = df_vert %>%
  slice(-which(df_vert$adult == df_vert$juvenile))
#84 with different vertical habitat use trait
```

Which species have the same or different horizontal habitat use traits?

```{r Check same horz traits, tidy=TRUE, message=FALSE}
df_horz = check_traits %>%
  select(prey_sp, life_stage, horz_habitat) %>%
  pivot_wider(., names_from = life_stage, values_from = horz_habitat)

check_traits_horz = df_horz %>%
  dplyr::slice(which(df_horz$adult == df_horz$juvenile))
#182 with same

check_traits_horz_not = df_horz %>%
  slice(-which(df_horz$adult == df_horz$juvenile))
#39 with different vertical habitat use trait
```

Which species have the same or different seasonal habitat use traits?

```{r Check same season traits, tidy=TRUE, message=FALSE}
df_season = check_traits %>%
  select(prey_sp, life_stage, season_migrant) %>%
  pivot_wider(., names_from = life_stage, values_from = season_migrant)

check_traits_season = df_season %>%
  dplyr::slice(which(df_season$adult == df_season$juvenile))
#189 with same

check_traits_season_not = df_season %>%
  slice(-which(df_season$adult == df_season$juvenile))
#32 with different vertical habitat use trait
```

Which species have the same or different diel vertical migration strategy traits?

```{r Check same diel traits, tidy=TRUE, message=FALSE}
df_diel = check_traits %>%
  select(prey_sp, life_stage, diel_migrant) %>%
  pivot_wider(., names_from = life_stage, values_from = diel_migrant)

check_traits_diel = df_diel %>%
  slice(which(df_diel$adult == df_diel$juvenile))
#182 with same

check_traits_diel_not = df_diel %>%
  slice(-which(df_diel$adult == df_diel$juvenile))
#39 with different vertical habitat use trait
```

There were 127 species for which trait information did not differ between adult and juvenile life stages.

```{r Join dataset with same traits, tidy=TRUE}
check_traits_same = check_traits %>%
  filter(prey_sp %in% check_traits_season$prey_sp) %>%
  filter(prey_sp %in% check_traits_diel$prey_sp) %>%
  filter(prey_sp %in% check_traits_horz$prey_sp) %>%
  filter(prey_sp %in% check_traits_vert$prey_sp)

length(unique(check_traits_same$prey_sp)) #127

length(unique(check_traits_same$prey_sp))*100/308
```

There were 94 species for which trait information differed in one or more traits for adult and juvenile life stages.

```{r check traits diff, tidy=TRUE}
check_traits_diff = check_traits %>%
  filter(!prey_sp %in% check_traits_same$prey_sp)
  #filter(!prey_sp %in% check_traits_diel$prey_sp) %>%
  #filter(!prey_sp %in% check_traits_season$prey_sp) %>%
  #filter(!prey_sp %in% check_traits_horz$prey_sp) %>%
  #filter(!prey_sp %in% check_traits_vert$prey_sp)
length(unique(check_traits_diff$prey_sp)) #94

length(unique(check_traits_diff$prey_sp))*100/308

write.csv(check_traits_diff, here::here("data/3_output_supp/check_traits_diff.csv"), row.names = FALSE)
```

Double check that these also add up to 221.

```{r Check some numbers, tidy=TRUE}
length(unique(check_traits_diff$prey_sp)) + length(unique(check_traits_same$prey_sp)) #Needs to add up to 221.
```

Now check on data proportions for albacore reported or measured and how much data this affects. To this we separate how many rows of data corresponded to "Estimated" or "Measured" albacore fork lengths. We then filter these data to match the species for which different trait information was on record but no further meta-information that would help establish the prey's likely life stage or trait information. We then report these outputs in the decision tree.

```{r Checking on }

check_studyID_df = data.frame(table(alb_global_diet$StudyID)) %>%
  mutate(prop = Freq*100/1027)

check_spp_alb_measured = alb_global_diet %>%
  filter(pred_flmean_est == "Measured: study reported") %>% #Note this is 391 rows of data out of 1027
  filter(prey_sp %in% check_traits_diff$prey_sp)
  
391*100/1027 #38% of the data
length(unique(check_spp_alb_measured$prey_sp)) #51 species with measured albacore lengths
length(unique(check_spp_alb_measured$prey_sp))*100/308 #16.5% of species with measured albacore lengths

check_spp_alb_estimated = alb_global_diet %>%
  filter(pred_flmean_est == "Estimated: by meta-analysis") %>% #Note this is 636 rows of data out of 1027
  filter(prey_sp %in% check_traits_diff$prey_sp) %>%
  filter(!prey_sp %in% check_spp_alb_measured$prey_sp)

636*100/1027 #62% of the data
length(unique(check_spp_alb_estimated$prey_sp)) #43 species with estimated albacore lengths
length(unique(check_spp_alb_estimated$prey_sp))*100/308 #14% of species with measured albacore lengths

length(unique(check_spp_alb_measured$prey_sp))+length(unique(check_spp_alb_estimated$prey_sp)) #Needs to sum to 94.
```

Edit decision tree with this information!

### Select traits for analyses

Key traits selection for species coarse traits: for habitat association, seasonal, vertical migration and aggregation behaviour.

```{r Prey traits selection, tidy=TRUE}

prey_trait_select = prey_trait %>%
  filter(!prey_sp %in% check_spp$prey_sp) %>%
  dplyr::select(prey_class:prey_sp, life_stage, diet_likely, life_note, 
                vert_habitat, horz_habitat, diel_migrant, diel_migrant_cat, refuge_cat, season_migrant, season_cat, gregarious_primary) %>%
  filter(life_stage == diet_likely)

prey_trait_select$prey_sp[duplicated(prey_trait_select$prey_sp)]

str(prey_trait_select)

```

This resulted in 93 adult, 210 juvenile and 5 larval prey taxa in the trait database.

```{r filter prey trait df by species remaining in diet df}

prey_trait_select_use = prey_trait_select %>%
  filter(prey_sp %in% unique(alb_global_metrics_basic$prey_sp)) %>%
  left_join(alb_global_metrics_basic)

summary(as.factor(prey_trait_select_use$life_stage))

write.csv(prey_trait_select_use, here::here("data/2_output_data/prey_trait_select.csv"), row.names = FALSE)

```

### Table S4 content - life stage and measurements

Include this in meta-information.

```{r Covariates, tidy = TRUE}

prey_trait_select_use = read.csv(here::here("data/2_output_data/prey_trait_select.csv"))

alb_table_S4 = alb_global %>%
  #filter(maxl_use > 0) %>%
  dplyr::select(prey_class:prey_sp, prey_age_reported_1, maxl_use, maxl_type, gape_lmax) %>%
  group_by(prey_sp, prey_age_reported_1, maxl_type) %>% #, life_stage , pFOuse, pN, pM maxl_use, , gape_lmax
  dplyr::summarize(`maxL (cm)` = max(maxl_use/10, na.rm=TRUE),
                   `maxGape (cm)` = max(gape_lmax, na.rm=TRUE)) %>%
  left_join(prey_trait_select_use, by="prey_sp") %>%
  select(-c(prey_class:prey_family, diet_likely, vert_habitat:maxM))

summary(as.factor(alb_table_S4$prey_age_reported_1))
summary(as.factor(alb_table_S4$life_stage))

write.csv(alb_table_S4, here::here("data/3_output_supp/table_S4.csv"), row.names = FALSE)
write.csv(alb_table_S4, here::here("outputs_figures/ms_figures/SI_results/table_S4.csv"), row.names = FALSE)
```

## Step 2 GO TO 2_extract_longhurst to work on Figure 1a
## Step 3 THEN TO document 3a_alb_data_covars to work on covariate info
## Step 4 TO 4_species_accum doc to work on Figure 1b and c
## Step 5 TO 5_trait_phylo_graph for species and trait phylogenetic graphs
## Step 6 TO 6a_trait_cluster_analysis_divis7 for clustering algorithm analyses & 6b_alb_diet_graph for trait relative frequency of occurrence graphics
## Step 7 TO 7a-c_trait_model code for trait-based modelling steps.
