==> code/2_extract_longhurst.Rmd <==
---
title: "Step 2 Geographic Data Processing"
author: "Natasha Hardy"
date: "17/03/2022"
output: html_document
---

## About

This document appends longhurst province code and information to our study location and covariate information, from the global albacore diet data project. 

## Workspace

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r load packages, message=FALSE}
#Spatial & plot
library(sf)
library(rnaturalearth)
library(ggplot2)
library(sp)
#Data
library(here)
"%notin%" = Negate('%in%')
here::here()
library(dplyr)

```

## Setting up the map

### Base map

```{r create base map, message=FALSE, tidy=TRUE}
# Base map
world_map <- rnaturalearth::ne_countries(scale = 'small', returnclass = c("sf"))
baseMap <- ggplot() +
  geom_sf(data = world_map, size = .2, fill = "gray80", col = "gray90") +
  theme(panel.grid.major = element_line(color = gray(0.9), linetype = "dashed", size = 0.5))
baseMap
```

### Add Longhurst Province Shapefile

```{r upload longhurst province shp}

# Load the Longhurst shapefile, convert to SpatialPolygonsDataFrame
longhurst <- st_read(here::here("code/longhurst/Longhurst_world_v4_2010.shp")) 
names(longhurst)
head(longhurst)
#Check data
plot(longhurst)
```

```{r plot longhurst province shp, message=FALSE}

# Plot provinces should work as before, just with annoying warning messages
sf_use_s2(FALSE)
longMap <- baseMap + geom_sf(data = longhurst, aes(fill = ProvCode), size = .2, col = "grey50", alpha=.4)+
  theme_bw() + theme(legend.position="none", axis.title = element_blank()) +
  geom_sf_text(data = longhurst %>% 
                 dplyr::group_by(ProvCode) %>% 
                 dplyr::summarize(n()), 
               aes(label = ProvCode),
               colour = "grey20", check_overlap = TRUE) +
  scale_fill_viridis_d()+ coord_sf(expand = FALSE)
longMap
```


### Manipulate Longhurst shp to spatial polygon df

```{r manip longhurst shapefile to spatial polygon df, message=FALSE}

longs <- sf:::as_Spatial(longhurst) # SpatialPolygonsDataFrame

st_crs(longs)
```

### Albacore data

```{r read in and manipulate albacore data, message=FALSE}

alb <- read.csv(here::here("data/2_output_data/alb_covars.csv"), head = TRUE, sep= ",")
rep_ID = alb$rep_ID
# Set coordinates
coordinates(alb) <- ~ LocatLongitude + LocatLatitude # SpatialPointsDataFrame
# Set the projection of the SpatialPointsDataFrames to match the projection of the shapefile
suppressWarnings(proj4string(alb) <- proj4string(longs))

# Extract the Longhurst province of each location
provinces <- over(alb, longs)
alb$province <- provinces$ProvDescr
alb$code <- provinces$ProvCode
provinces$rep_ID = rep_ID

write.csv(alb, here::here("data/2_output_data/alb_covars_longhurst.csv"), row.names = FALSE)
```

### Albacore data map
```{r make albacore study data map, message=FALSE, tidy=TRUE}
# Map
albMap <- longMap + 
  geom_point(data = as.data.frame(alb), aes(x = LocatLongitude, y = LocatLatitude)) +
  ylim(c(-55, 63))
albMap

ggsave(here::here("outputs_figures/1_maps/alb_global_map_f.pdf"), plot = albMap, dpi = 300)
ggsave(here::here("outputs_figures/1_maps/alb_global_map_f.jpeg"), plot = albMap, dpi = 300)
```

## Old Code

```{r Add Longhurst Province Shp, include=FALSE, eval=FALSE,message=FALSE, tidy=TRUE}
# Add longhurst regions information:
longhurst <- sf::read_sf(here::here("code/longhurst/Longhurst_world_v4_2010.shp")) 
# or wherever you have it stored

# simplify the object to make it 'usable'
# Gives a warning, one/some of the regions may be off?
sf::sf_use_s2(FALSE)
longhurst <- longhurst %>% 
  sf::st_simplify(dTolerance = 0.01) %>% 
  dplyr::group_by(ProvCode,ProvDescr) %>% 
  dplyr::summarise()

#Check data
plot(longhurst)
glimpse(longhurst)
```

### Note re troubles with sf package especially and comms with Barb

Re - warning messages and errors after loading the longhurst shapefile doc.

From Barb:

It's an issue with the update of sf (see here). You can use the fix they suggest, i.e. add "sf::sf_use_s2(FALSE)" before line 35. That works for me. 

You can just add "ylim(c(-52, 52)" or something similar to trim the map extent, in the plotting code. I would add "theme_bw()" just before "theme(legend.position ... ", as that will keep a border around the whole map area.

The warnings are annoying but they appear harmless. It's not uncommon for R to not like some features in imported shapefiles. The bounding box errors relate to province BPLR in the Arctic (if you subset out that row, you'll notice these warnings disappear). The other errors are mostly related to the fact that we're doing spatial things with unprojected polygons. That would be an issue if, say, we wanted to calculate the area of each polygon, which requires us to project the dataset into a coordinate reference system that uses meters rather than decimal degrees. But as we're just simplifying the shapefile and mapping it, I don't think it matters.
