==> code/4_species_accum.Rmd <==
---
title: "Step 4 Species Accumulation"
author: "Natasha Hardy"
date: "21/03/2022"
output: html_document
---
---
title: "Albacore Diet Synthesis Rarefaction"
author: "Natasha Hardy"
date: "20/10/2021"
output: html_document
---

## About

Illustrating the biological diversity in diets of a highly migratory predator, albacore tuna, across ocean basins and years sampled. We collected historical data on albacore tuna stomach contents, and plot the mean and cumulative species richness obtained with each additional data collection effort.

## Workspace

```{r setup, include=FALSE, echo=FALSE, message=FALSE}
knitr::opts_chunk$set(echo = TRUE, warning = FALSE)
```

```{r Load Workspace, message=FALSE}

#R documentation
library(pander)
library(formatR)

# general
library(tidyverse)
library(readxl)
library(readr)
library(plyr)
library(dplyr)
library(here)
"%notin%" = Negate('%in%')
here::here()

# Graphics
library(ggplot2)

# Biodiversity & analyses
library(BiodiversityR)
library(vegan)
library(tcltk)

#How many of these packages do we really need?
#library(viridis)
#library(RColorBrewer)
#library(wesanderson)
#library(devtools)

```

## Species community data

```{r Load in species community data, tidy=TRUE}
#Data -- using total diet data
alb_global_dietpa = read.csv(here("data/2_output_data/alb_global_wide_dietpa.csv"), 
          check.names = FALSE) %>%
  group_by(OceanBasin) %>%
  arrange(YearEnd)

## Albacore prey species P/A data
# data
PreyPAdata = as.data.frame(alb_global_dietpa[,10:ncol(alb_global_dietpa)])
PreyPAsite = as.data.frame(alb_global_dietpa[,c(1:5)])

#str(PreyPAdata)
str(PreyPAsite)

rownames(PreyPAsite) = rownames(PreyPAdata)

#check.datasets(PreyPAdata, PreyPAsite) 
#ensure that the sites of the community dataset are the same sites as those from the environmental dataset --> something that is assumed to be the case for the BiodiversityR and vegan packages.

#Save data if needed
#write.csv(PreyPAsite, here("data/output_data/alb_global_rarefactiondata.csv"), row.names = FALSE)

```

```{r Reload prey pa sites data, tidy=TRUE}

#PreyPAsite = read.csv(here("data/output_data/alb_global_rarefactiondata.csv"), check.names = FALSE)

PreyPAsite$YearEnd = as.integer(PreyPAsite$YearEnd)
PreyPAsite$StudyID = as.factor(PreyPAsite$StudyID)
PreyPAsite$OceanBasin = as.factor(PreyPAsite$OceanBasin)

```

Species richness per ocean basin:

```{r Summary species specnumber, tidy=TRUE}

# number of species and the factor that I want to group them by
SpO <- specnumber(PreyPAdata, PreyPAsite$OceanBasin)
SpO

```


## Mean Cumulative Richness

Species mean species accumulation overall:

Makes a dataset with species richness in relation to sites using the 'exact' method. This adds sites randomly, and finds the expected mean number of sites that contribute to a certain increase in species richness.

```{r Summary mean spec accum, tidy=TRUE}

Alb.mean.1 <- accumresult(PreyPAdata, y=PreyPAsite$OceanBasin, method='exact', conditioned=TRUE)
glimpse(Alb.mean.1)
accumplot(Alb.mean.1)

```

Mean cumulative richness in relation to Ocean Basin sampled:

```{r Site level mean spec accum, tidy=TRUE}
#With automatic plotting
Alb.mean.2 <- accumcomp(PreyPAdata, y=PreyPAsite, factor='OceanBasin', method='exact',
                        legend=FALSE, conditioned=TRUE) #, scale='YearEnd'
#Withhold on plotting so as to do so in ggplot
Alb.mean.3 <- accumcomp(PreyPAdata, y=PreyPAsite, factor='OceanBasin',
                     method='exact', conditioned=F, plotit=F)

Alb.mean.long3 <- accumcomp.long(Alb.mean.3, ci=NA, label.freq=5)

```

```{r Biodiversity R plotting theme by Roeland Kindt, tidy=TRUE}

# possibly need for extrafont::loadfonts(device="win") to have Arial
# as alternative, use library(ggThemeAssist)

BioR.theme <- theme(
  panel.background = element_blank(),
  #panel.border = element_blank(),
  panel.grid = element_blank(),
  axis.line = element_line("gray25"),
  text = element_text(size = 24, family="Arial"),
  axis.text = element_text(size = 24, colour = "gray25"),
  axis.title = element_text(size = 24, colour = "gray25"),
  legend.title = element_text(size = 24),
  legend.text = element_text(size = 24),
  legend.key = element_blank(),
  legend.position = c(0.8, 0.3)
  )

```

```{r Site level mean spec accum plot, tidy=TRUE}

plotgg1 <- ggplot(data=Alb.mean.long3, aes(x = Sites, y = Richness, ymax =  UPR, ymin= LWR)) + 
  scale_x_continuous(expand=c(0, 1), sec.axis = dup_axis(labels=NULL, name=NULL)) +
  scale_y_continuous(sec.axis = dup_axis(labels=NULL, name=NULL)) +
  #ylim(0,130)+
  geom_line(aes(colour=Grouping), size=2) +
  geom_point(data=subset(Alb.mean.long3, labelit==TRUE), 
             aes(colour=Grouping, shape=Grouping), size=5) +
  geom_ribbon(aes(colour=Grouping), alpha=0.2, show.legend=FALSE) + 
  BioR.theme +
  scale_color_brewer(palette = "Set1") +
  labs(x = "Cumulative sampling seasons", y = "Mean cumulative species richness", 
       colour = "Ocean Basin", shape = "Ocean Basin")

plotgg1

ggsave(here("outputs_figures/ms_figures/fig1_map_spp/alb_mean_accum_legend.jpeg"), plot=plotgg1, height = 8, width = 12, dpi = 300)

```

## Step-wise accumulation by Ocean Basin

Makes a dataset with species richness in relation to sites using the 'exact' method. This adds sites randomly, and finds the expected mean number of sites that contribute to a certain increase in species richness.

```{r Summary step-wise spec accum, tidy=TRUE}

Alb.acc.1 <- accumresult(PreyPAdata, y=PreyPAsite$OceanBasin, method='collector', conditioned=TRUE)
Alb.acc.1
accumplot(Alb.acc.1)
str(Alb.acc.1)

```

```{r Site level step-wise spec accum, tidy=TRUE}
#With automatic plotting
Alb.acc.2 <- accumcomp(PreyPAdata, y=PreyPAsite, factor='OceanBasin', method='collector',
                        legend=FALSE, conditioned=TRUE) #, scale='YearEnd'
#Withhold on plotting so as to do so in ggplot
Alb.acc.3 <- accumcomp(PreyPAdata, y=PreyPAsite, factor='OceanBasin',
                     method='collector', conditioned=F, plotit=F)

Alb.acc.long3 <- accumcomp.long(Alb.acc.3, ci=NA, label.freq=5)

```

```{r Biodiversity R plotting theme by Roeland Kindt 2nd theme, eval=FALSE, include=FALSE}
#One need to change this if necessary
# possibly need for extrafont::loadfonts(device="win") to have Arial
# as alternative, use library(ggThemeAssist)

BioR.theme2 <- theme(
  panel.background = element_blank(),
  #panel.border = element_blank(),
  panel.grid = element_blank(),
  axis.line = element_line("gray25"),
  text = element_text(size = 20, family="Arial"),
  axis.text = element_text(size = 20, colour = "gray25"),
  axis.title = element_text(size = 20, colour = "gray25"),
  legend.title = element_text(size = 20),
  legend.text = element_text(size = 20),
  legend.key = element_blank(),
  legend.position = c(0.8, 0.3))

```

```{r Site level step-wise spec accum plot, tidy=TRUE}

plotgg2 <- ggplot(data=Alb.acc.long3, aes(x = Sites, y = Richness, ymax =  UPR, ymin= LWR)) + 
  scale_x_continuous(expand=c(0, 1), sec.axis = dup_axis(labels=NULL, name=NULL)) +
  scale_y_continuous(sec.axis = dup_axis(labels=NULL, name=NULL)) +
  geom_line(aes(colour=Grouping), size=2) +
  geom_point(data=subset(Alb.acc.long3, labelit==TRUE), 
             aes(colour=Grouping, shape=Grouping), size=5) +
  BioR.theme +
  scale_color_brewer(palette = "Set1") +
  labs(x = "Cumulative sampling seasons", y = "Cumulative species richness", 
       colour = "Ocean Basin", shape = "Ocean Basin")

plotgg2

ggsave(here("outputs_figures/ms_figures/fig1_map_spp/alb_step_accum_simple.jpeg"), 
       plot=plotgg2, height = 8, width = 10, dpi = 300)

```


## Step-wise accumulation by Ocean Basin & Year sampled

```{r Manipulate and extract year data, eval=FALSE}

prey_year = PreyPAsite %>%
  arrange(desc(OceanBasin))

plot2_data = cbind(prey_year, Alb.acc.long3)

```

We need to retroactively append 'YearEnd' data to the species accumulation data obtained through Biodiversity R workflow. This code solution was suggest by Roeland Kindt. Both the above and this solution work, as the 'collector' method reads in the sites in the order they are sampled, in our case that is chronologically and nested within Ocean Basins.

```{r Alternative loop to extract year data for plot 3, tidy=TRUE}

Alb.acc.long3$Year <- NA
groupings <- unique(Alb.acc.long3$Grouping)
 
for(g in 1:length(groupings)) {
  group <- groupings[g]
  # assuming Grouping was based on variable Region of the env.data
  Alb.acc.long3[Alb.acc.long3$Grouping == group, "Year"] <- PreyPAsite[PreyPAsite$OceanBasin == group, "YearEnd"]
  }

```

```{r Species accum data manip if needed, eval=FALSE, include=FALSE}

#levels(as.factor(Alb.acc.long3$Grouping))

#Alb.acc.long3$Grouping <- factor(Alb.acc.long3$Grouping, levels = c("Mediterranean", "N Atlantic",
#                                                        "S Atlantic", "Indian",
#                                                        "N Pacific", "S Pacific"))

```

```{r Biodiversity R plotting theme by Roeland Kindt 3nd theme, tidy=TRUE}
#One need to change this if necessary
# possibly need for extrafont::loadfonts(device="win") to have Arial
# as alternative, use library(ggThemeAssist)

BioR.theme3 <- theme(
  panel.background = element_blank(),
  #panel.border = element_blank(),
  panel.grid = element_blank(),
  axis.line = element_line("gray25"),
  text = element_text(size = 24, family="Arial"),
  axis.text = element_text(size = 24, colour = "gray25"),
  axis.title = element_text(size = 24, colour = "gray25"),
  legend.title = element_text(size = 24),
  legend.text = element_text(size = 24),
  legend.key = element_blank(),
  legend.position ="none",
  #legend.position = c(0.2, 0.7)
  )

```

```{r Site level step-wise spec accum plot by year, tidy=TRUE}

plotgg3 <- ggplot(data=Alb.acc.long3, aes(x = Year, y = Richness, ymax =  UPR, ymin= LWR)) + 
  scale_x_continuous(expand=c(0, 1), sec.axis = dup_axis(labels=NULL, name=NULL)) +
  scale_y_continuous(sec.axis = dup_axis(labels=NULL, name=NULL)) +
  geom_line(aes(colour=Grouping), size=2) +
  geom_point(data=subset(Alb.acc.long3, labelit==TRUE), 
             aes(colour=Grouping, shape=Grouping), size=5) +
  BioR.theme3 +
  scale_color_brewer(palette = "Set1") +
  labs(x = "Years sampled", y = "Cumulative species richness", 
       colour = "Ocean Basin", shape = "Ocean Basin")

plotgg3

ggsave(here("outputs_figures/ms_figures/fig1_map_spp/alb_step_accum_year_nl.jpeg"), plot=plotgg3, height = 8, width = 12, dpi = 300)

```


## Extra code

If we want to relevel factors and make our own colour palettes. 

```{r rarefaction graphs - curve, eval=FALSE}

levels(as.factor(Alb.acc.long3$Grouping))

Alb.acc.long3$Grouping <- factor(Alb.acc.long3$Grouping, levels = c("Mediterranean", "N Atlantic",
                                                        "S Atlantic", "Indian",
                                                        "N Pacific", "S Pacific"))

Accum.plot.1 <- ggplot(data = Alb.acc.long3, aes(y=Richness, x=Year, colour=Grouping)) + #x=Sites
  geom_point(size=2) +
  geom_line(size=1.5) + 
  theme_classic() +
  theme(panel.grid.major = element_blank(), panel.grid.minor = element_blank()) +
  #scale_x_continuous(breaks = c(1,2,3,4,5,6,7,8,9,10,11,12,13,14,15))+
  theme(axis.title=element_text(size=18)) +
  theme(axis.text=element_text(size=18)) +
  theme(legend.text=element_text(size=18)) +
  theme(legend.title=element_text(size=18)) +
  labs(x="Years sampled", y="Species cumulative richness")+
  scale_colour_manual(values=wes_palette("Zissou1", 7, type="continuous"), name="Ocean Basin")+
  geom_errorbar(data = Alb.acc.long3, aes(ymin=Richness-SD, ymax=Richness+SD), width=.2)
Accum.plot.1

ggsave(here("outputs_figures/ms_figures/fig1_map_spp/accum.plot1.jpeg"), plot=Accum.plot, height = 8, width = 12, dpi = 300)
```

## Comms with Roeland Kindt

Dear Natasha,
 
Thank you very much for email and finding BiodiversityR useful.
 
To answer briefly (on holiday in Mombasa right now)
 
The accumcomp function does not include information on characteristics of sites. This is because typical methods randomize the sites. But indeed for the collector method, the first site is the first site that is encountered
Therefore indeed this information needs to be added, maybe using a script similar to the following after getting data first in the accumcomp.long format
 
accum.long1$Year <- NA
groupings <- unique(accum.long1$Grouping)
 
for (g in 1:length(groupings) {
group <- groupings[g]
Assuming Grouping was based on variable Region of the env.data
accum.long1[accum.long1$Grouping == group, “Year”] <- env.data[env.data$Region == group, “Year”]
}
 
Then it should be possible to use ggplot as in your example, with year as x variable. It should be a numeric variable to generate the type of plot that you made earlier.
 
Please come back to me in case I misunderstood something.
 
Best regards – stay safe,
 
Roeland
 
Roeland KINDT, dr. ir.
