==> code/6A_trait_cluster_analysis_divk7.Rmd <==
---
title: "Step 6 Prey Trait Cluster Analyses"
author: "Natasha Hardy"
date: "28/03/2022"
output: html_document
---

# Prey typologies identified in the diets of albacore tuna

## About

**Prey typologies text from Ms.** Here, we identified key prey typologies or functional groups in albacore tuna diets using hierarchical clustering calculated with divisive algorithms (Jain et al., 1999; Legendre & Legendre, 1998) on a Gower dissimilarity matrix (Gower, 1971) to identify relational structure between mixed ecological traits types: 3 binomial variables and 2 categorical variables for prey species traits (packages included: cluster, vegan and dendextend; code found under ‘Albacore_synthesis_b.Rmd’). Our objective in clustering selection (Brock et al., 2008; Charrad et al., 2014; Theodoridis & Koutroumbas, 2006) is to optimise the number of clusters based on (1) maximum differentiation or separation of species between clusters, (2) minimum differentiation of species or compactness within clusters, (3) optimal silhouette width coefficient value as well as Dunny Smith residuals, and (4) evenness or balance of cluster composition (number of species in each cluster). We assessed clusters visually for balance and consistency using cluster dendrograms and trait values that influenced a species’ occupancy within a cluster are visualised using heatmaps. The relative position of species to each other in relation to their cluster occupancy, and based on shared or separation of trait values, was visualised using multivariate ordination-based, non-metric multidimensional scaling (nMDS) (Field et al., 1982).

We use code and concepts described by Anastasia Reusova here: https://towardsdatascience.com/hierarchical-clustering-on-categorical-data-in-r-a27e578f2995

And cluster validation concepts further discussed here: https://www.datanovia.com/en/lessons/cluster-validation-statistics-must-know-methods/#silhouette-coefficient

**Note:** We excluded the aggregation behavioural trait from trait guild characterisation because the vast majority of prey species were schooling or aggregating taxa and we did not obtain congruent outputs when included. 

## Workspace

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, warning = FALSE)
```

```{r Worspace, results='hide', tidy=TRUE, message=FALSE}

#Need packages
library(plyr)
library(dplyr)
library(tidyverse)
library(reshape2)
library(factoextra)
library(here)
"%notin%" = Negate('%in%')
here::here()

#Markdown
library(formatR)

# Graphics
library(ggplot2)
library("PNWColors")
library(viridis)

#Multivariate work
library(vegan)
library(cluster)
library("dendextend")
library(NbClust)

```

## Prey traits load & cluster algorithms

### Prey & traits for cluster

Here we use the cleaned file for prey species traits based on probable life stage consumed by albacore. We need to select species for which we have complete trait information for the selected traits. We obtained 156 taxa in our previous version, but now 292 species due to filling in data.

Below we select variables for use in clustering algorithms. We selected vertical and horizontal habitat association, as well as diel and seasonal migratory traits, the probable life stage consumed and aggregation behaviour. We selected these traits because we hypothesised that these relate to the first level of filtering in the predation process -- encountering prey.

```{r Load and manip data for cluster analyses, tidy=TRUE, warning=FALSE, message=FALSE}

prey_traits_load = read.csv(here::here("data/2_output_data/prey_trait_select.csv"), header=TRUE) %>%   dplyr::select(-c(diel_migrant, refuge_cat, season_migrant, life_note, diet_likely)) %>% 
  dplyr::rename(gregarious = gregarious_primary) %>%
  mutate(diel_migrant_cat = case_when(diel_migrant_cat == "diel_no" ~ "diel (no)",
                                      diel_migrant_cat == "diel_UN" ~ "diel (unknown)",
                                      diel_migrant_cat == "diel_yes" ~ "diel (yes)"),
         `gregarious`=case_when(`gregarious` == "solitary" ~ "non-schooling",
                                `gregarious` == "pairing" ~ "non-schooling",
                                `gregarious` == "shoaling" ~ "schooling",
                                `gregarious` == "schooling" ~ "schooling"),
         season_cat = case_when(season_cat == "season_no" ~ "season (no)",
                                season_cat == "season_NA" ~ "season (unknown)",
                                season_cat == "season_yes" ~ "season (yes)")
         ) 
#Note that we are grouping shoaling and schooling as traits due to low numbers, as well as solitary and pairing species.

prey_traits = prey_traits_load %>%
  filter(diel_migrant_cat != "diel (unknown)",
         season_cat != "season (unknown)"
         ) %>%
  drop_na()

```

Check on distribution and levels of factors for cluster analyses.

```{r Additional data checks, tidy=TRUE, warning=FALSE}

summary(as.factor(prey_traits$life_stage)) #very few larva, mainly juve's and some adults
summary(as.factor(prey_traits$vert_habitat)) #consider merging benthic and demersal as we do later
#maybe also meso and bathypelagic?
summary(as.factor(prey_traits$horz_habitat)) #merge reef-associated with coastal, 
#as well as continental shelf and slope
summary(as.factor(prey_traits$diel_migrant_cat)) #already removed unknowns, most still DVM's
summary(as.factor(prey_traits$season_cat)) #already removed unknowns, most still seasonal
summary(as.factor(prey_traits$gregarious))

```

We will be combining benthic / demersal taxa due to function and low numbers, as well as meso and bathypelagic prey.

```{r Rename to manage low numbers}

#Levels vertical habitat
levels(as.factor(prey_traits$vert_habitat)) #check
prey_traits$vert_habitat <- factor(prey_traits$vert_habitat,
                                   levels = c("benthic", "demersal", "epipelagic",
                                              "mesopelagic", "bathypelagic"),
                                   labels = c("demersal", "demersal", "epipelagic",
                                              "mesopelagic", "mesopelagic"))
#Too low sample size for bathypelagic (merged with mesopelagic), and benthic merged with demersal

#Levels horizontal habitat
levels(as.factor(prey_traits$horz_habitat)) #check
prey_traits$horz_habitat <- factor(prey_traits$horz_habitat,
                                   levels = c("reef-associated", "coastal", "continental shelf",
                                              "continental slope", "oceanic"),
                                   labels = c("coastal", "coastal", "continental shelf", 
                                              "continental shelf", "oceanic"))

```

Manipulate dataframes for analyses:

```{r Data manip for cluster analyses and dendrogram, tidy=TRUE, warning=FALSE, message=FALSE}

## Row name as column datasets
#For probable prey traits
prey_traits_row <- prey_traits
dendlabs <- rownames(prey_traits_row)
prey_traits <- cbind(dendlabs, prey_traits_row)

#Probable trait df subsets
prey_species = prey_traits$prey_sp

#use_traits = as.data.frame(prey_traits[,c(7:12)]) #12
#probable_traits --> use_traits

use_traits = prey_traits %>%
  dplyr::select(vert_habitat, horz_habitat, diel_migrant_cat, season_cat) #life_stage, , gregarious
use_traits[,1:4] = lapply(use_traits[,1:4], as.factor)

```

The full trait dataframe including quantitative data to append to cluster number:

```{r prey traits df}
str(prey_traits)
```

The subset trait dataframe to run clustering analysis

```{r traits for cluster analysis df}
str(use_traits)
```

### Cluster techniques

Here we generate a multivariate species-based dissimilarity matrix of the data, and a hierarchical divisive clustering algorithm in order to select the most appropriate algorithm.

```{r Cluster and Dissimilarity Matrix, warning=FALSE, message=FALSE}

#### Distance measure
#select just the traits you want to contribute to ordination
prob.gower.dist <- daisy(use_traits, metric = c("gower"))

#### Divisive cluster
prob.divisive.clust <- diana(as.matrix(prob.gower.dist), 
                              diss = TRUE, keep.diss = TRUE)
plot(prob.divisive.clust, main = "Divisive")

```

### Cluster Assessment Output

**Run cluster stats table function**

```{r Cluster stat table function, warning=FALSE, message=FALSE}

# Cluster stats comes out as list while it is more convenient to look at it as a table
# This code below will produce a dataframe with observations in columns and variables in row
# Not quite tidy data, which will require a tweak for plotting, but I prefer this view as an output here as I find it more comprehensive 
library(fpc)
cstats.table <- function(dist, tree, k) {
  clust.assess <- c("cluster.number","n","within.cluster.ss","average.within","average.between",
                    "wb.ratio","dunn2","avg.silwidth")
  clust.size <- c("cluster.size")
  stats.names <- c()
  row.clust <- c()
  output.stats <- matrix(ncol = k, nrow = length(clust.assess))
  cluster.sizes <- matrix(ncol = k, nrow = k)
  for(i in c(1:k)){
    row.clust[i] <- paste("Cluster-", i, " size")
  }
  for(i in c(2:k)){
    stats.names[i] <- paste("Test", i-1)
    
    for(j in seq_along(clust.assess)){
      output.stats[j, i] <- unlist(cluster.stats(d = dist, clustering = cutree(tree, k = i))[clust.assess])[j]
      
    }
    
    for(d in 1:k) {
      cluster.sizes[d, i] <- unlist(cluster.stats(d = dist, clustering = cutree(tree, k = i))[clust.size])[d]
      dim(cluster.sizes[d, i]) <- c(length(cluster.sizes[i]), 1)
      cluster.sizes[d, i]
      
    }
  }
  output.stats.df <- data.frame(output.stats)
  cluster.sizes <- data.frame(cluster.sizes)
  cluster.sizes[is.na(cluster.sizes)] <- 0
  rows.all <- c(clust.assess, row.clust)
  # rownames(output.stats.df) <- clust.assess
  output <- rbind(output.stats.df, cluster.sizes)[ ,-1]
  colnames(output) <- stats.names[2:k]
  rownames(output) <- rows.all
  is.num <- sapply(output, is.numeric)
  output[is.num] <- lapply(output[is.num], round, 2)
  output
}

```

**Notes:** 20-24/07/2020 and updated 28/03/2022

Ultimately we are aiming for distinct clusters of species, such that the difference within clusters is minimal and between clusters is maximised. Assessing cluster statistical tables, we are consistently observing lower average.within cluster differences using agglomerative clustering compared to divisive algorithms.

Below we are using simplified habitat use trait values + gregarious (as binary) traits for: divisive clustering techniques. We assess primarily: 

(i) Balance between and within clusters == the number of species per cluster and between cluster. We are looking for the method which provides the greatest balance.

(ii) Balance of the lowest 'average.within' and greatest 'average.between' differences between clusters.

(iii) Lower 'dunn2' or dunny smith residual values.

(iv) Higher 'avg.sildwidth' or average silhouette width values.

**NOTE: We observe the greatest balance between these cluster validation criteria for the hierarchical divisive clustering algorithm and for k = 7 clusters.**

```{r Cluster Assessment Tables, warning=FALSE, message=FALSE}

#Stats table for divisive method
prob.stats.df.divisive <- cstats.table(prob.gower.dist, prob.divisive.clust, 15)
prob.stats.df.divisive
View(prob.stats.df.divisive)
write.csv(prob.stats.df.divisive, here("outputs_figures/3_clusters/prob.stats.df.divisive.hab.csv"))

```

**Note** Divisive -- for visualising cluster number selection.

* Elbow method (29/03/2022) using habitat use + gregarious (binary) + life stage use inflection at 6 (and 7)

* Silhouette (29/03/2022) using habitat use + gregarious (binary) + life stage use inflection at 6 (and 7)

* Other metrics such as average between, cluster balance, and dunny smith residuals support k = 6 or 7.

* Will try 7 for this method.

```{r Cluster Selection - Divisive, warning=FALSE, message=FALSE}

### Using "Elbow" and "Silhouette" methods to identify the best number of clusters

# Elbow method

prob.div.elbow = ggplot(data = data.frame(t(cstats.table(prob.gower.dist, prob.divisive.clust, 20))), 
       aes(x=cluster.number, y=within.cluster.ss)) + 
  geom_point()+
  geom_line()+
  ggtitle("Divisive clustering") +
  labs(x = "Num.of clusters", y = "Within clusters sum of squares (SS)") +
  theme(plot.title = element_text(hjust = 0.5))
prob.div.elbow
ggsave(here('outputs_figures/3_clusters/prob.div.elbow.hab.png'), 
       plot=prob.div.elbow, width=8, height=8, dpi=300)

## Silhouette
#When it comes to silhouette assessment, the rule is you should choose the number that maximizes the 
#silhouette coefficient because you want clusters that are distinctive (far) enough to be considered separate.

prob.div.silh = ggplot(data = data.frame(t(cstats.table(prob.gower.dist, prob.divisive.clust, 20))), 
       aes(x=cluster.number, y=avg.silwidth)) + 
  geom_point()+
  geom_line()+
  ggtitle("Divisive clustering") +
  labs(x = "Num.of clusters", y = "Average silhouette width") +
  theme(plot.title = element_text(hjust = 0.5))
prob.div.silh
ggsave(here('outputs_figures/3_clusters/prob.div.silh.hab.png'), 
       plot=prob.div.silh, width=8, height=8, dpi=300)

```

## DIVISIVE k = 7

### Cluster Dendrograms

Using k = 6 for the habitat association, aggregation behaviour & probable life stage traits.

**Horizontal dendrogram**

```{r Horizontal dendrogram, warning=FALSE, message=FALSE}
prob.dendro <- as.dendrogram(prob.divisive.clust)
PNW.pal7 <- pnw_palette(7, name = "Bay", type = "continuous")
###Horizontal dendrogram - Probable traits

#Horizontal cluster illustration version
prob.dendro.col <- prob.dendro %>%
  set("branches_k_color", k = 7, value = PNW.pal7) %>%
  set("branches_lwd", 0.8) %>%
  set("labels", prey_species) %>% #NOT VERY LEGIBLE...
  set("labels_colors", 
      value = c("darkslategray")) %>% 
  set("labels_cex", 0.5)
prob.ggd1 <- as.ggdend(prob.dendro.col)
prob.dendro.graph <- ggplot(prob.ggd1, theme = theme_minimal()) +
  labs(x = "Num. observations", y = "Height", title = "Dendrogram, k = 6")
prob.dendro.graph

ggsave(here('outputs_figures/3_clusters/prob.dendro.horz.k7.png'), plot=prob.dendro.graph, width=8, height=8, dpi=300)

```

**Radial dendrogram**

```{r Radial dendrogram, warning=FALSE, message=FALSE}

# Radial plot looks less cluttered (and cooler)
prob.dendro.rad <- ggplot(prob.ggd1, labels = FALSE) + 
  scale_y_reverse(expand = c(0.2, 0)) +
  coord_polar(theta="x")
prob.dendro.rad 
#No labels on this one, labels were too cluttered/problems
#Save radial dendrogram for chat

ggsave(here('outputs_figures/3_clusters/prob.dendro.rad.k7.png'), plot=prob.dendro.rad, width=8, height=8, dpi=300)

```

**Vertical dendrogram**

Similar to https://stackoverflow.com/questions/38034663/rotate-labels-for-ggplot-dendrogram 

```{r Vertical dendrogram, warning=FALSE, message=FALSE}

# This is a different way to compute hierarchical clustering and cut the tree
#clus <- hcut(mydist, k = 6, hc_func = 'hclust', hc_method = 'ward.D2', graph = FALSE, isdiss = TRUE)

#Below is problematic0
#labels(dend) <- paste0(paste0(rep('', 3), collapse = ''), speciesO)
#dend <- sort(dend, decreasing = FALSE)
#View(labels(dend))

#Creating df for the dend labels so that we can accurately line them up with the species
dendlabs <- labels(prob.dendro) #Need to create strings of labels to manipulate
dendlabs2 <- as.data.frame(dendlabs) #turn in df

#Join these data so we can relabel the dendrogram
#Use plyr function because it conserves the row order of the left df, which matters for assigning labels here

dfdend <- plyr::join(dendlabs2, prey_traits)

ggd1 <- ggplot(prob.dendro %>%
                 set("branches_k_color", k = 7, value = PNW.pal7) %>%
                 set("branches_lwd", 0.8) %>%
                 set("labels", dfdend$prey_sp) %>% #NOT VERY LEGIBLE...
                 set("labels_colors", 
                     value = c("darkslategray")) %>% 
                 set("labels_cex", 0.5), 
               theme = theme_minimal(),
               horiz = TRUE)

ggd1 <- ggd1 + theme(panel.grid.major = element_blank(),
                     axis.text = element_blank(),
                     axis.title = element_blank())
ggd1 <- ggd1 + ylim(max(get_branches_heights(prob.dendro)), -1)
ggd1

ggsave(here('outputs_figures/3_clusters/prob.dendro.vertlabs.k7.pdf'), plot=ggd1, width=5, height=20, dpi=300)

#With labels removed!!!

prob.dendro.vert <- ggplot(prob.ggd1, horiz = TRUE, labels = FALSE) + 
  scale_y_reverse(expand = c(0.2, 0)) #+
#coord_polar(theta="x")
prob.dendro.vert 
#Export as .png
ggsave(here('outputs_figures/3_clusters/prob.dendro.vert2.k7.png'), plot=prob.dendro.vert, width=5, height=12, dpi=300)


```

### Cluster Heatmaps 

```{r Extract cluster number to trait matrix, warning=FALSE, message=FALSE}

#Extract cluster number to trait matrix
prob.clust.num <- cutree(prob.divisive.clust, k = 7)

#we want to bind the original dataset with the cluster numbers such that each species is assigned a cluster
#can use whole data or just traits use to just look at unique species clusters in relation to traits
#alb.cl <- cbind(ctraitsO, alb.clust.num)
#OR
prob.prey.cl <- cbind(prey_traits, prob.clust.num)

#View(prob.prey.cl)

write.csv(prob.prey.cl, here("outputs_figures/3_clusters/prob.prey.clusternum.divis.k7.csv"))
write.csv(prob.prey.cl, here("data/2_output_data/prob.prey.clusternum.divis.k7.csv"))

```

```{r load cluster data back in}


prob.prey.cl = read.csv(here::here("data/2_output_data/prob.prey.clusternum.divis.k7.csv")) %>%
  select(-c(X))

```

```{r Cluster heatmap dfs, warning=FALSE, message=FALSE}
# Time for the heatmap
# the 1st step here is to have 1 variable per row
# factors have to be converted to characters in order not to be dropped

#Note plyr can mess with this!!
detach("package:plyr", unload=TRUE)

#Create dfs for graphs
prob.clust.long = prob.prey.cl %>%
  dplyr::select(prey_sp, vert_habitat:season_cat, `prob.clust.num`) %>% #maxFO:maxM, 
  rename(Vertical = vert_habitat,
         Horizontal = horz_habitat,
         `Diel Migrant` = diel_migrant_cat,
         `Seaonsal Migrant` = season_cat) %>%
  reshape2::melt(id.vars = c("prey_sp", "prob.clust.num"), variable.name = "trait", value.name = "level") %>%
  group_by(prob.clust.num, trait, level) %>%
  mutate(count = n_distinct(prey_sp)) %>%
  distinct(prob.clust.num, trait, level, count) %>% #, percent
  group_by(prob.clust.num, trait) %>%
  mutate(percent = count / sum(count)*100) %>%
  arrange(prob.clust.num)

View(prob.clust.long)

```

```{r Cluster heatmap, warning=FALSE, message=FALSE}
#heatmap.c will be suitable in case you want to go for absolute counts - but it doesn't tell much to my taste
#problem below involves the values of our data being ordinal, therefore they are not unique
levels(prob.clust.long$trait)
levels(as.factor(prob.clust.long$prob.clust.num))

#Our data above comes truncated, you would need to truncate the data and re-label clusters depending on which dfs you melt/merge/reshape.
#Example: View(alb.cust.long.q[96:nrow(alb.cust.long.q),])
heatmap.c <- ggplot(prob.clust.long, aes(x = factor(prob.clust.num), y = level)) +
  geom_tile(aes(fill = count))+
  theme(panel.grid.major = element_blank(),
        axis.title=element_text(size=24),
        axis.text=element_text(size=24),
        legend.text=element_text(size=24),
        legend.title=element_text(size=24)) +
  labs(x = "Trait guild (cluster number)", y = NULL) +
  scale_fill_viridis(option="magma", begin = 0.2, end = 0.95)+
  facet_grid(trait~. , scales="free_y")+
  theme(strip.text = element_text(face="bold", size=rel(2)))
heatmap.c

ggsave(here('outputs_figures/3_clusters/prob.dendro.heatcounts.divis.k7.clean.pdf'), plot=heatmap.c, width=12, height=12, dpi=300)
ggsave(here('outputs_figures/3_clusters/prob.dendro.heatcounts.divis.k7.clean.jpeg'), plot=heatmap.c, width=12, height=12, dpi=300)


heatmap.p <- ggplot(prob.clust.long, aes(x = factor(prob.clust.num), y = factor(level, ordered = T))) +
  geom_tile(aes(fill = percent), alpha = 0.85)+
   theme(panel.grid.major = element_blank(),
        axis.title=element_text(size=24),
        axis.text=element_text(size=24),
        legend.text=element_text(size=24),
        legend.title=element_text(size=24)) +
  labs(x = "Trait guild (cluster number)", y = NULL) +
  #labs(title = "Distribution of characteristics across clusters by percentage", x = "Cluster number", y = NULL) +
  #scale_fill_gradient2(low = "darkslategray1", mid = "yellow", high = "turquoise4") +
  scale_fill_viridis(option="magma", begin = 0.2, end = 0.95)+
  facet_grid(trait~., scales="free_y")+
  theme(strip.text = element_text(face="bold", size=rel(2)))
heatmap.p

ggsave(here('outputs_figures/3_clusters/prob.dendro.heatpercent.divis.k7.clean.pdf'), plot=heatmap.p, width=12, height=12, dpi=300)
ggsave(here('outputs_figures/3_clusters/prob.dendro.heatpercent.divis.k7.clean.jpeg'), plot=heatmap.p, width=12, height=12, dpi=300)

library(plyr)

```

```{r Summary info, warning=FALSE, message=FALSE}
summary(as.factor(prob.prey.cl$prob.clust.num))
#1  2  3  4  5  6  7 
#12 88 51 14 55 44 28 

```

## NMDS for checking on our ordination - divisive

Here we want to visualise species' occupancy of trait-based cluster in multivariate space. Species have been treated as sites and their trait occupancy as 'species' in the ordination and nMDS routines. Thus their position in multivariate space is based off similar and dissimilar trait values. We then overlay their cluster number graphically.

### Ordination

```{r nMDS dissimilarity, tidy=TRUE, warning=FALSE, message=FALSE, results='hide'}

#prob.gower.dist <- daisy(use_traits, metric = c("gower"))
#str(prob.gower.dist)
#### nMDS for dissimilarity
trait_NMDS_prob <- metaMDS(prob.gower.dist #, 
                           #k = 2, trymax = 100
                           ) 
#solution after 20 iterations #prob.gower.dist
trait_NMDS_prob[["stress"]] #stress = 0.1477077 #reasonable
plot(trait_NMDS_prob )

```

### Extract NMDS coordinates and associate with co-variates/grouping factors

```{r Data scores for plotting, warning=FALSE, message=FALSE}

#Extract NMDS coordinates and associate with co-variates/grouping factors
data.scores = as.data.frame(trait_NMDS_prob$points)

#Using the scores function from vegan to extract the site scores and convert to a data.frame
#data.scores <- as.data.frame(scores(trait_NMDS_prob))  #, "species"
#?scores()
#create a column of site names, from the rownames of data.scores
data.scores$points <- rownames(data.scores)

#bind treatment labels and score values
treatment.scores <- cbind(prob.prey.cl, data.scores)
unique(treatment.scores$prob.clust.num)

```

### NMDS plot

This is unfortunately not very informative as the species are very clumped by their trait values. We could figure out how to 'jitter' the points such that we have clumps?

```{r make manual palette to match the cluster colours}

library(scales)

fig3.pal <- c("#00496F", "#EDD746", 
              "#0A718F", "#59A082", 
              "#EDA417", "#E7720B", "#DD4124")
#Edit this!!
#fig3.pal <- c("#00496F", "#94B669", "#0A718F", 
#              "#EDA417", "#EDB829", 
#              "#E7720B", "#DD4124")

show_col(fig3.pal)

```


```{r nMDS Cluster Number, warning=FALSE, message=TRUE}
#nMDS plot for the assessmenet of global change (yes/no) and the drivers.
#as.integer(unique(treatment.scores$adult.clust.num))
#treatment.scores$prob.clust.num = lapply(treatment.scores$prob.clust.num, as.factor)

trait_nMDS_prob_fig <- ggplot() + 
  #geom_polygon(data=hull.data, 
  #             aes(x=NMDS1, y=NMDS2, 
  #                 fill= prob.clust.num, 
  #                 group= prob.clust.num), 
  #             alpha=0.30) + # add the convex hulls
  geom_point(data=treatment.scores, 
             aes(x=MDS1, y=MDS2, colour=factor(prob.clust.num)), 
             size=2,
             position = position_jitter(width = 0, height = 0.1, seed = 8)) + # add the point markers
  #geom_jitter(aes(x=NMDS1, y=NMDS2, colour=prob.clust.num))+
  coord_equal() +
  theme_bw()  +
  theme(axis.text.x = element_blank(),  # remove x-axis text
        axis.text.y = element_blank(), # remove y-axis text
        axis.ticks = element_blank(),  # remove axis ticks
        axis.title.x = element_text(size=18), # remove x-axis labels
        axis.title.y = element_text(size=18), # remove y-axis labels
        legend.title = element_text(size = 18), 
        legend.text = element_text(size = 18),
        legend.justification = c(0,0.5),
        #panel.background = element_rect(fill = "lightgrey"), 
        panel.grid.major = element_blank(),  #remove major-grid labels
        panel.grid.minor = element_blank(),  #remove minor-grid labels
        plot.background = element_blank()) +
  scale_colour_manual(values=fig3.pal, name="Cluster")
  
#  scale_colour_viridis(option="magma", begin = 0.8, end = 0.2, name = "Cluster Number") #+ 
  #scale_fill_viridis(option="magma", begin = 0.8, end = 0.2, name = "Cluster Number") #+
  
trait_nMDS_prob_fig

#Note plyr can mess with this!!
detach("package:plyr", unload=TRUE)

ggsave(here::here("outputs_figures/3_clusters/trait_nMDS_prob_clusters_div7_b.png"), plot = trait_nMDS_prob_fig, width = 8, height = 8, dpi = 300)

```
