==> code/6B_alb_diet_graph.Rmd <==
---
title: "Albacore Diet Synthesis Graphs"
author: "Natasha Hardy"
date: '2022-06-07'
output: html_document
---

## About

To produce graphical illustrations of global albacore diet composition in relation to prey clusters/guilds and relative importance of prey trait guilds.

## Workspace

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r Packages to load}
library(here)
"%notin%" = Negate('%in%')
here::here()
library(tidyverse)
library(readxl)
library(readr)
library(plyr)
library(dplyr)
library(vegan)
library(mvabund)
library(ggplot2)
library(viridis)
library(RColorBrewer)
library(ggh4x)
library(scales)

#library(wesanderson)

```

## Load Data

```{r Load frequency of occurrence data, tidy=TRUE}

#diet_fo_wide <- read.csv(here("./data/output_data/alb_global_diet_wide_dietfo.csv"))

alb_global_diet = read.csv(here::here("data/2_output_data/alb_global_diet_total.csv"), check.names = FALSE)

prob.prey.cl = read.csv(here::here("data/2_output_data/prob.prey.clusternum.divis.k7.csv")) %>%
  select(-c(X, dendlabs, maxFO:maxM))

```

```{r Combine diet and cluster info, tidy=TRUE}

alb_diet_clust = alb_global_diet %>%
  left_join(prob.prey.cl) %>%
  #filter(prob.clust.num != "NA") %>% #> 0 #, pFOuse > 5
  select(StudyID, rep_ID, YearEnd, OceanBasin, OceanBasinQ, prey_class:prey_sp, pFOuse, pN, pM, prob.clust.num)
#unique(alb_diet_clust$prob.clust.num)
#unique(alb_diet_clust$OceanBasin)

alb_diet_clust$YearEnd = as.integer(alb_diet_clust$YearEnd)
alb_diet_clust$prob.clust.num = as.factor(alb_diet_clust$prob.clust.num)

unique(alb_diet_clust$StudyID) #26
```

## Index of relative importance for FO

Here we will calculate a normalised metric or index for diet composition, of relative contribution to the total frequency of occurrence of all species’ within each trait guild, normalised for each replicate diet observation.

### DF for FO index of relative importance

```{r Create df for range graph, tidy=TRUE}

index_alb_diet_clust = alb_diet_clust %>%
  group_by(StudyID, rep_ID, YearEnd, OceanBasin, prob.clust.num) %>%
  dplyr::summarize(clustFO = sum(pFOuse, na.rm=TRUE)) 

totals_alb_diet_clust = alb_diet_clust %>%
  group_by(StudyID, rep_ID, YearEnd, OceanBasin) %>%
  dplyr::summarize(totalFO = sum(pFOuse, na.rm=TRUE)) 

index_alb_diet_clust_data = index_alb_diet_clust %>%
  left_join(totals_alb_diet_clust) %>%
  mutate(relFO = clustFO/totalFO) %>%
  dplyr::filter(relFO > 0)

index_alb_diet_clust_data$yearrepID <- paste(index_alb_diet_clust_data$YearEnd, index_alb_diet_clust_data$rep_ID, sep = "_")

write.csv(index_alb_diet_clust_data, here::here("data/2_output_data/index_alb_diet_clust_data.csv"))
```

```{r Load index data, tidy=TRUE}

index_alb_diet_clust_data = read.csv(here::here("data/2_output_data/index_alb_diet_clust_data.csv")) %>%
  select(-X)

length(unique(index_alb_diet_clust_data$StudyID)) #studies containing useable FO data
length(unique(index_alb_diet_clust_data$rep_ID)) #observations of useable FO data
```

### Colour palette

Match the colour palette for clusters in these and other graphs in the MS.

```{r Colour palette for cluster graphs, tidy=TRUE}
unique(index_alb_diet_clust_data$prob.clust.num)
fig3.pal <- c("#00496F", "#EDD746", 
              "#0A718F", "#59A082", 
              "#EDA417", "#E7720B", "#DD4124",
              "#787878")
scales::show_col(fig3.pal)
```

### Figure for index data - stacked bar

```{r Try stacked bar chart}

index_cluster_bar = index_alb_diet_clust_data %>%
  ggplot() +
  geom_bar(data = index_alb_diet_clust_data, 
             aes(y=relFO, x= yearrepID, #interaction(YearEnd, rep_ID), #yearrepID, #group=rep_ID, YearEnd:rep_ID, 
                 fill=as.factor(prob.clust.num)),
             width = 1, stat = "identity") + #, position = "jitter", , position="identity"
  #geom_jitter(position = position_jitter(width = 0.1, height = 0.1))+
  theme(panel.grid.major = element_blank(),
        axis.title=element_text(size=35),
        axis.title.x=element_blank(),
        axis.text=element_text(size=30),
        #axis.text.x = element_blank(),#
        axis.text.x = element_text(angle = 70, vjust = 1, hjust=1),
        legend.text=element_text(size=35),
        legend.title=element_text(size=35),
        #legend.position = "top"
        legend.position = c(.20, .85),
        legend.background = element_rect(fill="lightgrey")
        )+
  labs(x="Years sampled", y="Relative %FO")+
  scale_x_discrete(labels= c("1880 (Jordan, 1880)", 
  "1928 (Bouxin, 1934-40)", "1929 (Bouxin, 1934-40)", "1930 (Bouxin, 1934-40)", "1931 (Bouxin, 1934-40)", "1932  (Bouxin, 1934-40)", "1933  (Bouxin, 1934-40)", 
  "1941 (Hart, 1948)", "1941 (Hart, 1948)", "1945 (Hart, 1948)", "1946 (Hart, 1948)", "1946 (Hart, 1948)", "1947 (Hart, 1948)", "1949 (Hart, 1948)",
  "1957 (Iversen, 1962)", "1964 (Matthews, 1977)", "1968 (Ortiz Zarate, 1987)", "1968 (Pinkas, 1971)", "1968 (Pinkas, 1971)", "1968 (Pinkas, 1971)",
  "1969 (Pinkas, 1971)", "1969 (Pinkas, 1971)", "1971 (Aloncle, 1973)", "1983 (Bernard, 1985)", "1992 (Bello, 1999)", "1993 (Pusineri, 2005)", "1994 (Bello, 1999)",
  "1998 (Dos Santos, 2002)", "2001 (Watanabe, 2004)", "2002 (Watanabe, 2004)", "2002 (Watanabe, 2004)", "2004 (Goni, 2011)", "2005 (Goni, 2011)",
  "2005 (Goni, 2011)", "2005 (Goni, 2011)", "2005 (Goni, 2011)", "2005 (Goni, 2011)", "2006 (Consoli, 2008)", "2006 (Glaser, 2015)", "2006 (Goni, 2011)",
  "2006 (Goni, 2011)", "2006 (Goni, 2011)", "2006 (Goni, 2011)", "2006 (Goni, 2011)", "2006 (Young, 2010)", "2006 (Young, 2010)", "2007 (Goni, 2011)",
  "2007 (Salman, 2009)", "2008 (Goni, 2011)", "2008 (Romeo, 2012)", "2008 (Williams, 2015)", "2009 (Williams, 2015)", "2010 (Madigan, 2015)",
  "2010 (Williams, 2015)", "2010 (Williams, 2015)", "2010 (Williams, 2015)", "2014 (Romanov, 2020)", "2014 (Romanov, 2020)", "2014 (Romanov, 2020)",
  "2015 (Romanov, 2020)")) + #unique(index_alb_diet_clust_data$StudyID)
  scale_fill_manual(values=fig3.pal, name="Prey Trait Guilds",
                    labels = c("1 (non-diel migrating mesopelagics)", 
                               "2 (diel migrating mesopelagics)", 
                               "3 (coastal and shelf epipelagics)",
                               "4 (coastal and shelf demersals)", 
                               "5 (oceanic epipelagics)", 
                               "6 (seasonal continental shelf)",
                               "7 (resident continental shelf)", 
                               "NA (not classified)")) +
  facet_grid(OceanBasin~.)+ #, scales ="free", StudyID, scales ="free_x"
  theme(strip.text = element_text(face="bold", size=rel(4)))
index_cluster_bar

#unique(index_alb_diet_clust_data$YearEnd)
#unique(index_alb_diet_clust_data$StudyID)
#levels(as.factor(index_alb_diet_clust_data$yearrepID))

#ggsave(here::here("outputs_figures/4_diet_comp/index_cluster_barFO.jpeg"), plot = index_cluster_bar, width = 42, height = 33, dpi = 300)
#ggsave(here::here("outputs_figures/ms_figures/fig4_dietcomp/index_cluster_barFO.jpeg"), plot = index_cluster_bar, width = 42, height = 33, dpi = 300)

ggsave(here::here("outputs_figures/4_diet_comp/index_cluster_barFO_b.jpeg"), plot = index_cluster_bar, width = 34, height = 31, dpi = 150)
ggsave(here::here("outputs_figures/ms_figures/fig4_dietcomp/index_cluster_barFO_b.jpeg"), plot = index_cluster_bar, width = 34, height = 31, dpi = 150)

ggsave(here::here("outputs_figures/4_diet_comp/index_cluster_barFO_b2.jpeg"), plot = index_cluster_bar, width = 48, height = 32, dpi = 300)
ggsave(here::here("outputs_figures/ms_figures/fig4_dietcomp/index_cluster_barFO_b2.jpeg"), plot = index_cluster_bar, width = 48, height = 32, dpi = 300)
```

### Figure for index data - scatter - extra

```{r Try scatterplot relative FO, eval=FALSE}

index_cluster_scatter = index_alb_diet_clust_data %>%
  #filter(pFOuse > 5) %>%
  ggplot() +
  geom_point(data = index_alb_diet_clust_data, 
             aes(y=relFO, x=YearEnd, group=prob.clust.num, fill=prob.clust.num),
             size=5, shape=23, position = "dodge"
             ) +
  geom_jitter(position = position_jitter(width = 0.1, height = 0.1))+
  theme(panel.grid.major = element_blank(),
        axis.title=element_text(size=24),
        axis.text=element_text(size=24),
        legend.text=element_text(size=24),
        legend.title=element_text(size=24),
        legend.position = "top") +
  labs(x="Years sampled", y="Range in frequency of occurrence (%)")+
  scale_fill_manual(values=fig3.pal, name="Prey Trait Guilds") +
  #ylim(0,90)+
  xlim(1900,2019)+
  facet_grid(OceanBasin~.)+ #, scales ="free"
  #force_panelsizes(cols = c(0.6, 0.85, 1, 1, 0.4, 0.4),
  #                 rows = c(0.6, 0.6, 0.6, 1, 1, 0.6, 0.6)) +
  theme(strip.text = element_text(face="bold", size=rel(2)))
index_cluster_scatter

ggsave(here::here("outputs_figures/4_diet_comp/diet_cluster_scatterFO.jpeg"), plot = index_cluster_scatter, width = 24, height = 16, dpi = 300)

```

## Total data + manip for diet composition

#### Scatterplots

```{r Try scatterplot pFO, message=FALSE}

diet_cluster_scatter = alb_diet_clust %>%
  filter(pFOuse > 5) %>%
  ggplot() +
  geom_point(data = alb_diet_clust, aes(y=pFOuse, x=YearEnd,
                                        group=prob.clust.num,
                                        fill=prob.clust.num),
             
             size=5, shape=23) +
  coord_flip()+
  theme(panel.grid.major = element_blank(),
        axis.title=element_text(size=24),
        axis.text=element_text(size=24),
        legend.text=element_text(size=24),
        legend.title=element_text(size=24),
        legend.position = "none") +
  labs(x="Years sampled", y="Range in frequency of occurrence (%)")+
  scale_fill_manual(values=fig3.pal) + #, name="Prey Trait Guilds"
  #scale_fill_brewer(palette="BrBG", name="Prey Trait Guilds")+
  ylim(0,90)+
  xlim(1900,2019)+
  facet_grid(prob.clust.num~OceanBasin, scales ="free")+
  force_panelsizes(cols = c(0.6, 0.85, 1, 1, 0.4, 0.4),
                   rows = c(0.6, 0.6, 0.6, 1, 1, 0.6, 0.6)) +
  theme(strip.text = element_text(face="bold", size=rel(2)))
diet_cluster_scatter

ggsave(here::here("outputs_figures/4_diet_comp/diet_cluster_scatterFO.jpeg"), plot = diet_cluster_scatter, width = 24, height = 16, dpi = 300)

```

```{r Try scatterplot pN, message=FALSE}

diet_cluster_scatterN = alb_diet_clust %>%
  filter(pN > 5) %>%
  ggplot() +
  geom_point(data = alb_diet_clust, aes(y=pN, x=YearEnd,
                                        group=prob.clust.num,
                                        fill=prob.clust.num),
             
             size=5, shape=23) +
  coord_flip()+
  theme(panel.grid.major = element_blank(),
        axis.title=element_text(size=24),
        axis.text=element_text(size=24),
        legend.text=element_text(size=24),
        legend.title=element_text(size=24),
        legend.position = "none") +
  labs(x="Years sampled", y="Range in numerical abundance (%)")+
  scale_fill_manual(values=fig3.pal) + #, name="Prey Trait Guilds"
  #scale_fill_brewer(palette="BrBG", name="Prey Trait Guilds")+
  ylim(0,90)+
  xlim(1900,2019)+
  facet_grid(prob.clust.num~OceanBasin, scales ="free")+
  theme(strip.text = element_text(face="bold", size=rel(2)))
diet_cluster_scatterN

ggsave(here::here("outputs_figures/4_diet_comp/diet_cluster_scatterN.jpeg"), plot = diet_cluster_scatterN, width = 24, height = 16, dpi = 300)

```

```{r Try scatterplot pM, message=FALSE}

diet_cluster_scatterM = alb_diet_clust %>%
  filter(pN > 5) %>%
  ggplot() +
  geom_point(data = alb_diet_clust, aes(y=pM, x=YearEnd,
                                        group=prob.clust.num,
                                        fill=prob.clust.num),
             size=5, shape=23) +
  coord_flip()+
  theme(panel.grid.major = element_blank(),
        axis.title=element_text(size=24),
        axis.text=element_text(size=24),
        legend.text=element_text(size=24),
        legend.title=element_text(size=24),
        legend.position = "none") +
  labs(x="Years sampled", y="Range in biomass contribution (%)")+
  scale_fill_manual(values=fig3.pal) + #, name="Prey Trait Guilds"
  #scale_fill_brewer(palette="BrBG", name="Prey Trait Guilds")+
  ylim(0,100)+
  xlim(1900,2019)+
  facet_grid(prob.clust.num~OceanBasin, scales ="free")+
  theme(strip.text = element_text(face="bold", size=rel(2)))
diet_cluster_scatterM

ggsave(here::here("outputs_figures/4_diet_comp/diet_cluster_scatterM.jpeg"), plot = diet_cluster_scatterM, width = 24, height = 16, dpi = 300)

```

#### Boxplots

```{r Try boxplot FO, message=FALSE}

diet_cluster_box = alb_diet_clust %>%
  filter(pFOuse > 5) %>%
  ggplot() +
  geom_boxplot(data = alb_diet_clust, 
                      aes(y=pFOuse, x=YearEnd, 
                          group=prob.clust.num, 
                          fill=prob.clust.num
                          ),
              position = "dodge") + #fill #YearEnd 
  #geom_violin()+
  coord_flip()+
  theme(panel.grid.major = element_blank(),
        axis.title=element_text(size=24),
        axis.text=element_text(size=24),
        legend.text=element_text(size=24),
        legend.title=element_text(size=24),
        legend.position = "none") +
  labs(x="Years sampled", y="Range in frequency of occurrence (%)")+
  scale_fill_manual(values=fig3.pal) + #, name="Prey Trait Guilds"
  #scale_fill_brewer(palette="BrBG", name="Prey Trait Guilds")+
  #ylim(0,90)+
  #xlim(1900,2019)+
  facet_grid(prob.clust.num~OceanBasin, scales ="free")+
  force_panelsizes(cols = c(0.6, 0.85, 1, 1, 0.4, 0.4),
                   rows = c(0.6, 0.6, 0.6, 1, 1, 0.6, 0.6)) +
  theme(strip.text = element_text(face="bold", size=rel(2)))
diet_cluster_box

ggsave(here::here("outputs_figures/4_diet_comp/diet_cluster_boxFO.jpeg"), plot = diet_cluster_box, width = 24, height = 16, dpi = 300)

```

#### Violin plots

```{r Try violin FO, message=FALSE}

diet_cluster_violin = alb_diet_clust %>%
  filter(pFOuse > 5) %>%
  ggplot() +
  geom_violin(data = alb_diet_clust, 
                      aes(y=pFOuse, x=YearEnd, 
                          group=prob.clust.num, 
                          fill=prob.clust.num
                          ),
              position = "dodge") + #fill #YearEnd 
  #geom_violin()+
  coord_flip()+
  theme(panel.grid.major = element_blank(),
        axis.title=element_text(size=24),
        axis.text=element_text(size=24),
        legend.text=element_text(size=24),
        legend.title=element_text(size=24),
        legend.position = "none") +
  labs(x="Years sampled", y="Range in frequency of occurrence (%)")+
  scale_fill_manual(values=fig3.pal) + #, name="Prey Trait Guilds"
  #scale_fill_brewer(palette="BrBG", name="Prey Trait Guilds")+
  #ylim(0,90)+
  #xlim(1900,2019)+
  facet_grid(prob.clust.num~OceanBasin, scales ="free")+
  force_panelsizes(cols = c(0.6, 0.85, 1, 1, 0.5, 0.4),
                   rows = c(0.6, 0.6, 0.6, 1, 1, 0.6, 0.6)) +
  theme(strip.text = element_text(face="bold", size=rel(2)))
diet_cluster_violin

ggsave(here::here("outputs_figures/4_diet_comp/diet_cluster_violinFO.jpeg"), plot = diet_cluster_violin, width = 24, height = 16, dpi = 300)

```

```{r Try violin and scatter FO, message=FALSE}

diet_cluster_violin_scatter = alb_diet_clust %>%
  filter(pFOuse > 5) %>%
  ggplot() +
  geom_point(data = alb_diet_clust, aes(y=pFOuse, x=YearEnd,
                                        group=prob.clust.num,
                                        colour=prob.clust.num),
             size=5, shape=23) +
  geom_violin(data = alb_diet_clust, 
                      aes(y=pFOuse, x=YearEnd, 
                          group=prob.clust.num, 
                          fill=prob.clust.num
                          ),
              position = "dodge") + #fill #YearEnd 
  #geom_violin()+
  coord_flip()+
  theme(panel.grid.major = element_blank(),
        axis.title=element_text(size=24),
        axis.text=element_text(size=24),
        legend.text=element_text(size=24),
        legend.title=element_text(size=24),
        legend.position = "none") +
  labs(x="Years sampled", y="Range in frequency of occurrence (%)")+
  scale_fill_manual(values=fig3.pal) + #, name="Prey Trait Guilds"
  scale_colour_manual(values=fig3.pal) + #, name="Prey Trait Guilds"
  #ylim(0,90)+
  #xlim(1900,2019)+
  facet_grid(prob.clust.num~OceanBasin, scales ="free")+
  force_panelsizes(cols = c(0.6, 0.85, 1, 1, 0.5, 0.4),
                   rows = c(0.6, 0.6, 0.6, 1, 1, 0.6, 0.6)) +
  theme(strip.text = element_text(face="bold", size=rel(2)))
diet_cluster_violin_scatter

ggsave(here::here("outputs_figures/4_diet_comp/diet_cluster_violin_scatterFO.jpeg"), plot = diet_cluster_violin_scatter, width = 24, height = 16, dpi = 300)


```
