==> code/7a_trait_model_data_sub3removed.Rmd <==
---
title: "Albacore Diet Synthesis C"
author: "Natasha Hardy"
date: "15/04/2021"
output:
  html_document: default
  pdf_document: default
---

# Trait variation in albacore diet across geographies

## About

This document contains code for data manipulation and refinement, to pipe to multivariate community analyses and traitglm on albacore diets across ocean basins, using habitat association, gregariousness traits and cluster # output from cluster analyses.

## Blueprint for paragraphs:

1. (Results) Describe the dominant significant traits and non-sig traits in relation to albacore age and geography sampled (Fig. R3).

2. Relate to clusters (potential additional analysis/figure).

3. Discussion - Investigate how species traits may be used to explain variation in prey use by albacore tuna using a multi-matrix approach.

## Current Issues / Limitations to discuss in paper

- We did use community data that included single occurrence species, this meant species that only occurred once in albacore diets across the world. As these are also associated with trait information and we aim to use traits to explain differences in communities, we were interested in using this information. 

- In a traditional multivariate analysis, single occurrence species would inflate differences between communities that are analysed on a species-only basis, whilst our sampling (a.k.a. tuna foraging) does not adequately pick up these rarer species. It is possible that the 'composition=TRUE' argument in traitglm() would likely assist in accounting for species, it does not constitute a full accounting of potential random variance from species low abundances.

- We therefore removed 206 species with low prevalence (occurs < 3 times), under code chunk 4.

- The data are longer (57 obs) than they are wide (98 species).

- Decision not to use phylogenetic order information, not enough replication at the individual order level, could define a phylogeny as in PGLMM example code below and that actually accounts for phylogenetic structure and relatedness rather than treating phylogenetic information as a trait: https://daijiang.github.io/phyr/reference/pglmm_compare.html 

## Workspace

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

#library("rstudioapi")
#setwd(dirname(getActiveDocumentContext()$path))
```

```{r Workspace, echo=FALSE, message=FALSE, results='hide'}
#If need to clear 'Environment'
#First view environment if you like
#ls()
#To clear your environment
#remove(list = ls())

# general
library(readxl)
library(plyr)
library(dplyr)
library(devtools)
library(reshape2)
library(here)
"%notin%" = Negate('%in%')
here::here()
library(pander)
library(stringr)

# graphics
library(ggplot2)
library(lattice)
library(graphics)
library(RColorBrewer)
library(gridExtra)
library(captioner)
library(viridis)
library("grid")
library("ggplotify")

#modelling
library(Hmisc)
library(mgcv)
library(unmarked)
library(lme4)
library(mvabund)
library(tweedie)
library(vegan)

## mvabund bug source code
#source("code/traitglm.R")
#source("code/get_polys.R")

```

## Data (L)

Load presence/absence data. Originally 58 observations and 316 for N Atlantic, N Pacific and Mediterranean. We exclude the Indian, S Atlantic and S Pacific from further analyses due to low replication (n < 7).

```{r Species PA (L) load data for oceans, tidy=TRUE, warning=FALSE, message=FALSE}

#Upload + manipulate Species & Covar DF
#Doesn't like the here() function now for some reason
alb_prey_pa = read.csv(paste(here::here("data/2_output_data/alb_global_wide_dietpa.csv"),
                                   sep=""), check.names=FALSE) %>% #, row.names = 1
  dplyr::select(StudyID:ncol(.), -`Nyctiphanes australis`) %>% #remove species outlier
  filter(OceanBasin %notin% c("Indian", "S Atlantic", "S Pacific")) #not enough replication for these values

print(dim(alb_prey_pa)) #58 316 from 70 obs after removing the other ocean basins noted above

summary(as.factor(alb_prey_pa$OceanBasin))

```

Now we need to check for and remove any zero-sum columns or rows of data, and low FO data and adjust the co-variate data.

```{r Species PA (L) remove low abundance taxa for oceans, tidy=TRUE, warning=FALSE, message=FALSE}
#Check for empty columns
which(colSums(alb_prey_pa[,10:ncol(alb_prey_pa)])<2) #Remove
length(which(colSums(alb_prey_pa[,10:ncol(alb_prey_pa)])<3)) #206 species have low prevalence
names(which(colSums(alb_prey_pa[,10:ncol(alb_prey_pa)])<2))

#alb_prey_use = alb_prey_pa[colnames(alb_prey_pa) %notin% names(which(colSums(alb_prey_pa[,10:ncol(alb_prey_pa)])<2))]

alb_prey_use = alb_prey_pa[colnames(alb_prey_pa) %notin% names(which(colSums(alb_prey_pa[,10:ncol(alb_prey_pa)])<3))]

#Check for empty rows
which(rowSums(alb_prey_use[,10:ncol(alb_prey_use)])==0) #None

#If so --> remove
#alb_prey_use = alb_prey_use %>%
#  slice(-c(60,61))

```

This results in 58 observations and 134 species that occur >1 time.

```{r Select species only (L) for oceans, tidy=TRUE, warning=FALSE, message=FALSE}
#Subset species df
alb_prey_use_spp = alb_prey_use[,10:ncol(alb_prey_use)]

names(alb_prey_use_spp) <- str_replace_all(names(alb_prey_use_spp), " ", "_")

#check df
dim(alb_prey_use_spp) #98 species & 57 observations

```

## Check (L) multivariate data

```{r Visualise nMDS ordinations PA, eval = FALSE, include = FALSE}


#PA
check_nmds = metaMDS(alb_prey_use_spp,
                     distance="bray",
                     trymax = 20,#10000, #1000
                     k=3) #could increase to k = 4

#saveRDS(alb_pa_ord, here("outputs_figures/traitglms/alb_prob_nMDS_pa.rds"))

#Note pb with this!! insufficient data?

ordiplot(check_nmds, type = "none") #creates blank plot
orditorp(check_nmds, display = "species", col = "black", air=0.01) #labels traits
orditorp(check_nmds, display = "sites", col = "red", air=0.01) #labels traits

```

## Traits (Q)

Using output of the cluster analyses, this included 292 species and 10 trait variables + meta-information.

```{r Trait DF (Q) inc clust num for oceans, tidy=TRUE, warning=FALSE, message=FALSE}

alb_prob_clust = read.csv(paste(here::here("./data/2_output_data/prob.prey.clusternum.divis.k7.csv"), 
                                sep=""), header=T, row.names = 1) %>%
   dplyr::select(prey_sp:prob.clust.num)

dim(alb_prob_clust) #we obtain 292 species, 11 traits or covariates
```

Previously we used order information, there are multiple issues with using this information at this stage. Firstly, we lack replication at the level of the phylogenetic order used here, secondly we treated phylogenetic information as a trait similar to a fixed factor. There are better ways to assign variance to phylogeny, including through phylogenetic structure, related, branch distances, and many studies recommend using phylogeny as a random variable that accounts for latent variance attributed to phylogenetic structure and evolution within a community rather than as a fixed factor.

```{r Traits with phylogenetic order info}

add_order = read.csv(paste(here::here("./data/2_output_data/prey_trait_select.csv"))) %>%
  select(prey_sp, prey_order)

alb_trait_use = alb_prob_clust %>%
  left_join(add_order) %>%
  select(prey_sp, prey_order, life_stage, vert_habitat, horz_habitat,
         diel_migrant_cat, season_cat, gregarious, prob.clust.num)

```

Here, we filter out low occurrences using 'alb_prob_pa' df. Note that we obtain 129 (< 134) species and their traits, so we'll need to filter the species df by the traits.

```{r Manip trait DF (Q) for oceans, tidy=TRUE, warning=FALSE, message=FALSE}

#alb_trait_use = alb_prob_clust %>%
#  select(prey_sp, life_stage, vert_habitat, horz_habitat,
#         diel_migrant_cat, season_cat, gregarious, prob.clust.num)

#Relabel to add underscore in names
alb_trait_use$prey_sp = str_replace_all(alb_trait_use$prey_sp, " ", "_")
#Assign species as row names
rownames(alb_trait_use) <- alb_trait_use$prey_sp
#Order species list
alb_trait_use <- alb_trait_use[ order(row.names(alb_trait_use)), ]

#Ensure traits are treated as factors for now
alb_trait_use[,4:9] <- lapply(alb_trait_use[,4:9], factor) #to convert to factors

#Need to filter by prey species we want to use.
alb_trait_use = alb_trait_use %>%
  filter(prey_sp %in% names(alb_prey_use_spp))
  #we retain 292 species and 8 trait values

dim(alb_trait_use)
str(alb_trait_use)
```

**Trait value checks**

Here we check the trait values for each trait, and check the distribution of values. Note we merge certain values with low occurrences of species within them, and if we have a logical reason to merge.

```{r Trait values check for oceans, warning=FALSE, message=FALSE}
summary(alb_trait_use$life_stage) #low numbers for larva, do not use

summary(alb_trait_use$vert_habitat) #use

summary(alb_trait_use$horz_habitat) #use

summary(alb_trait_use$diel_migrant_cat) #a lot more diel migrants than not

summary(alb_trait_use$season_cat) #more seasonal taxa than not

summary(alb_trait_use$gregarious) #still more gregarious than not, just an artefact of the pelagic prey data.

summary(as.factor(alb_trait_use$prob.clust.num)) #good distribution of species within clusters

#summary(as.factor(alb_trait_use$prey_order)) #an unaccounted artefact of the pelagic prey data.
```

**Cleaning and preparing trait values for analyses**

## Data (L) -- Select

Here we filter our prey species (presence/absence data) by the species that were also used for cluster analysis, because we have complete trait information for these species, as well as frequency of occurrence data.

```{r FILTER PREY BY PROB TRAITS for ocean basins, tidy=TRUE, warning=FALSE, message=FALSE}

alb_prey_use <- alb_prey_use_spp[colnames(alb_prey_use_spp) %in% rownames(alb_trait_use)]

dim(alb_prey_use)
```

## Covars (R) - MANIPS

Checking the level replication for each covariate. Note that our best covars ==> pred_life, OceanBasin, lat_cat, and maybe code (Longhurst) - but would need to exclude low numbered levels.

```{r Select oceans covars}

alb_covars = alb_prey_pa %>%
  select(StudyID:pred_life_adjust) #%>%
  #slice(-c(60,61))

#As factors
alb_covars[,1:ncol(alb_covars)] =  lapply(alb_covars[,1:ncol(alb_covars)], as.factor)

```

```{r Covars oceans (R) checking, eval=FALSE}
#Check representation
#summary(alb_adult_covars)

summary(alb_covars$lat_cat) #don't use
#temperate  tropical 
#     56         1 

summary(alb_covars$pred_life) #don't use
#  adult juvenile    mixed 
#   1       37       19

alb_covars$pred_life = factor(alb_covars$pred_life, levels = c("juvenile", "adult", "mixed"),
                              labels = c("juvenile", "mixed", "mixed"))

summary(alb_covars$OceanBasin)
#   Mediterranean    N Atlantic     N Pacific     ##S Pacific  #don't use S Pacific
#         9            21            28             ##5 
#Remove Indian and S Atlantic Oceans from analyses

summary(alb_covars$OceanBasinQ)
#Not going to use

summary(alb_covars$code)
#Not going to use
#Build an analysis for CCAL, MEDI, NADR!!

dim(alb_covars) #58 observations for 9 covars
unique(alb_covars$StudyID)
```

## DF checks

```{r Check names and df dimensions for ocean basins, eval=FALSE}

## Make sure species names match abund matrix
rownames(alb_trait_use) == colnames(alb_prey_use) #awesome!

#Check all matrix dimensions
dim(alb_trait_use) #129 species and 9 traits
dim(alb_prey_use) #58 obs, 129 species
dim(alb_covars) #58 obs, 9 covariates

```

## Data (L) - for Longhurst analysis

Load presence/absence data. Originally 58 observations and 316 for N Atlantic, N Pacific and Mediterranean. We exclude the Indian, S Atlantic and S Pacific from further analyses due to low replication (n < 7).

```{r Species PA (L) load data filter for longhurst use, tidy=TRUE, warning=FALSE, message=FALSE}

#Upload + manipulate Species & Covar DF
#Doesn't like the here() function now for some reason
alb_prey_longhurst = read.csv(paste(here::here("data/2_output_data/alb_global_wide_dietpa.csv"),
                                   sep=""), check.names=FALSE) %>% #, row.names = 1
  dplyr::select(StudyID:ncol(.), -`Nyctiphanes australis`) %>% #remove species outlier
  filter(OceanBasin %notin% c("Indian", "S Atlantic", "S Pacific"),
         code %in% c("CCAL", "NADR", "MEDI")) #not enough replication for these values
  

print(dim(alb_prey_longhurst)) #49 316 from 70 obs after removing the other ocean basins noted above

```

Now we need to check for and remove any zero-sum columns or rows of data, and low FO data and adjust the co-variate data.

```{r Species PA (L) remove low abundance taxa for longhurst use, tidy=TRUE, warning=FALSE, message=FALSE}
#Check for empty columns
which(colSums(alb_prey_longhurst[,10:ncol(alb_prey_longhurst)])<2) #Remove

length(which(colSums(alb_prey_longhurst[,10:ncol(alb_prey_longhurst)])<2)) #181 species have low prevalence

names(which(colSums(alb_prey_longhurst[,10:ncol(alb_prey_longhurst)])<2))

alb_prey_use_longhurst = alb_prey_longhurst[colnames(alb_prey_longhurst) %notin% names(which(colSums(alb_prey_longhurst[,10:ncol(alb_prey_longhurst)])<2))]

#Check for empty rows
which(rowSums(alb_prey_use_longhurst[,10:ncol(alb_prey_use_longhurst)])==0) #None

#If so --> remove
#alb_prey_use = alb_prey_use %>%
#  slice(-c(60,61))

```

This results in 49 observations and 126 species that occur >1 time.

```{r Select species only (L) for longhurst use, tidy=TRUE, warning=FALSE, message=FALSE}
#Subset species df
alb_prey_use_longhurst_spp = alb_prey_use_longhurst[,10:ncol(alb_prey_use_longhurst)]

names(alb_prey_use_longhurst_spp) <- str_replace_all(names(alb_prey_use_longhurst_spp), " ", "_")

#check df
dim(alb_prey_use_longhurst_spp) #134 species & 58 observations

```

## Traits (Q) - for Longhurst analysis

Here, we filter out low occurrences using 'alb_prob_pa' df and select species used for Longhurst analyses. Note that we obtain 129 (< 134) species and their traits, so we'll need to filter the species df by the traits.

```{r Manip trait DF (Q) for longhurst use, tidy=TRUE, warning=FALSE, message=FALSE}
#Relabel to add underscore in names
alb_prob_clust$prey_sp = str_replace_all(alb_prob_clust$prey_sp, " ", "_")
#Assign species as row names
rownames(alb_prob_clust) <- alb_prob_clust$prey_sp
#Order species list
alb_prob_clust <- alb_trait_use[ order(row.names(alb_prob_clust)), ]

alb_trait_use_longhurst = alb_prob_clust %>%
  select(prey_sp, life_stage, vert_habitat, horz_habitat,
         diel_migrant_cat, season_cat, gregarious, prob.clust.num) %>%
  filter(prey_sp %in% names(alb_prey_use_longhurst_spp))

#Ensure traits are treated as factors for now
alb_trait_use_longhurst[,2:8] <- lapply(alb_trait_use_longhurst[,2:8], factor) #to convert to factors

dim(alb_trait_use_longhurst) #121 species, 8 trait variables
str(alb_trait_use_longhurst)

```

**Trait value checks**

Here we check the trait values for each trait, and check the distribution of values. Note we merge certain values with low occurrences of species within them, and if we have a logical reason to merge.

```{r Trait values check for longhurst use, warning=FALSE, message=FALSE}
summary(alb_trait_use_longhurst$life_stage) #low numbers for larva, do not use

summary(alb_trait_use_longhurst$vert_habitat) #use

summary(alb_trait_use_longhurst$horz_habitat) #use

summary(alb_trait_use_longhurst$diel_migrant_cat) #a lot more diel migrants than not

summary(alb_trait_use_longhurst$season_cat) #more seasonal taxa than not

summary(alb_trait_use_longhurst$gregarious) #still more gregarious than not, just an artefact of the pelagic prey data.

summary(alb_trait_use_longhurst$prob.clust.num) #good distribution of species within clusters

#summary(as.factor(alb_trait_use$prey_order)) #an unaccounted artefact of the pelagic prey data.
```

## Data (L) -- Select

Here we filter our prey species (presence/absence data) by the species that were also used for cluster analysis, because we have complete trait information for these species, as well as frequency of occurrence data.

```{r FILTER PREY BY PROB Longhurst data, tidy=TRUE, warning=FALSE, message=FALSE}

alb_prey_use_longhurst <- alb_prey_use_longhurst_spp[colnames(alb_prey_use_longhurst_spp) %in% rownames(alb_trait_use_longhurst)]

dim(alb_prey_use_longhurst) #49 rows and 121 spp

```

## Covars (R) - MANIPS

Checking the level replication for each covariate. Note that our best covars ==> pred_life, OceanBasin, lat_cat, and maybe code (Longhurst) - but would need to exclude low numbered levels.

```{r Select longhurst covars}

alb_covars_longhurst = alb_prey_longhurst %>%
  select(StudyID:pred_life_adjust) #%>%
  #slice(-c(60,61))

#As factors
alb_covars_longhurst[,1:ncol(alb_covars_longhurst)] =  lapply(alb_covars_longhurst[,1:ncol(alb_covars_longhurst)], as.factor)

```

```{r Covars longhurst (R) checking, eval=FALSE}
#Check representation
#summary(alb_adult_covars)

summary(alb_covars_longhurst$lat_cat) #don't use
#temperate 
#     49   

summary(alb_covars_longhurst$pred_life) #don't use
#  juvenile  mixed 
#    33       16

summary(alb_covars_longhurst$OceanBasin)
#   Mediterranean    N Atlantic     N Pacific 
#         9            16            24       
#Remove Indian and S Atlantic Oceans from analyses

summary(alb_covars_longhurst$code)
#Build an analysis for CCAL, MEDI, NADR!!
#CCAL MEDI NADR 
#  24    9   16

dim(alb_covars) #49 observations for 9 covars

```


## SAVE DFS FOR ANALYSES

Save for Ocean Basin analysis.

```{r Saving ocean dfs for analysis}

#L - species df
write.csv(alb_prey_use, here::here("./data/2_output_data/alb_prey_pa_use.csv"), row.names=FALSE)

#R - covars df
write.csv(alb_covars, here::here("./data/2_output_data/alb_covars_use.csv"), row.names=FALSE)

#Q - trait df
write.csv(alb_trait_use, here::here("./data/2_output_data/alb_trait_use.csv"), row.names=FALSE)

```

Save for Longhurst Province Code analysis

```{r Saving longhurst dfs for analysis}

#L - species df
write.csv(alb_prey_use_longhurst, here::here("./data/2_output_data/alb_prey_pa_use_longhurst.csv"), row.names=FALSE)

#R - covars df
write.csv(alb_covars_longhurst, here::here("./data/2_output_data/alb_covars_longhurst.csv"), row.names=FALSE)

#Q - trait df
write.csv(alb_trait_use_longhurst, here::here("./data/2_output_data/alb_trait_use_longhurst.csv"), row.names=FALSE)

```
