==> code/7B_trait_model_analysis.Rmd <==
---
title: "Albacore Diet Synthesis C Trait Modelling"
author: "Natasha Hardy"
date: "15/04/2021"
output:
  html_document: default
  pdf_document: default
---

# Trait variation in albacore diet across geographies

## About

This document contains code for loading mostly manipulated data for global albacore prey communities consumed and for conducting multivariate community analyses and traitglm on albacore diets across ocean basins, using habitat association, gregariousness traits and cluster # output from cluster analyses. Note that current analyses in this document includes the following distribution of observations:

Mediterranean    N Atlantic     N Pacific
       9            21            28      

## Current work

Troubleshoot following analysis: BM suggested (email from 3/06/2020):
You might also see differences between upwelling systems (e.g. California Current, Canary Current) vs. downwelling systems (e.g. Agulhas Current, both coasts of Australia), but that should also fall out in the ocean basin analysis, given the sample distribution


## Workspace

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

#library("rstudioapi")
#setwd(dirname(getActiveDocumentContext()$path))
```

```{r Workspace, echo=FALSE, message=FALSE, results='hide'}
# general
library(tidyverse)
library(dplyr)
library(devtools)
library(here)
"%notin%" = Negate('%in%')
here::here()

# graphics
library(ggplot2)
library(lattice)
library(graphics)
library(ggplotify)
#modelling
library(mvabund)

## mvabund bug source code
source(here::here("code/traitglm.R"))
source(here::here("code/get_polys.R"))
```

## Data needs

For Ocean Basin analysis:

- Species (L): alb_prey_pa_use ==> "./data/output_data/alb_prey_pa_use.csv"

- Env/Site (R): alb_covars_use ==> "./data/output_data/alb_covars_use.csv"

- Traits (Q): alb_trait_use ==> "./data/output_data/alb_trait_use.csv"

Can also use either of the above for life stage of predator assessment + report results to Desiree.

## Load DFs for analyses

```{r Load Prey data (L)}

#L - species df
alb_prey_pa_use = read.csv(here::here("./data/2_output_data/alb_prey_pa_use.csv"))
dim(alb_prey_pa_use)

```

```{r Load Covar data (R)}

#R - covars df
alb_covars_use = read.csv(here::here("./data/2_output_data/alb_covars_use.csv"))

#As factors
alb_covars_use[,1:ncol(alb_covars_use)] =  lapply(alb_covars_use[,1:ncol(alb_covars_use)], as.factor)

```

```{r Load Trait data (Q)}

#Q - trait df
alb_trait_use = read.csv(here::here("./data/2_output_data/alb_trait_use.csv"))

#Needs species as rows
#Assign speces as rownames
rownames(alb_trait_use) <- alb_trait_use$prey_sp
#Order species list if needed
alb_trait_use <- alb_trait_use[ order(row.names(alb_trait_use)), ]
#Traits as factors for analyses
alb_trait_use[,1:ncol(alb_trait_use)] =  lapply(alb_trait_use[,1:ncol(alb_trait_use)], as.factor)

dim(alb_trait_use)

```

## DF Checks

**Covars**

```{r Covars values check, eval=FALSE}

#Summary stats on covars
unique(alb_covars_use$StudyID) #22 studies
unique(alb_covars_use$OceanBasin) #3 Ocean basins

```

## DF manips for multivariate analyses

```{r Check names and df dimensions}

## Make sure species names match abund matrix
rownames(alb_trait_use) == colnames(alb_prey_pa_use) #awesome!

# Then create the mvabund matrix
alb_prey_pa_use_mv = as.mvabund(alb_prey_pa_use)

#Check again that species names match trait values
rownames(alb_trait_use) == colnames(alb_prey_pa_use_mv) #awesome!

#Check all matrix dimensions
dim(alb_trait_use) #good
dim(alb_prey_pa_use_mv) #good
dim(alb_covars_use) #good

```

```{r Select covars traits and levels}

#Traits - select and manipulate that df - ideally we want to use the same df here
traits_use = alb_trait_use[,c(4:7,9)] %>%
  dplyr::rename(cluster = prob.clust.num) 

str(traits_use)
```

## TRAIT GLMS - MANYGLMS

### Read in models

```{r Read in manyglm models and anovas}
#Manyglm
ocean_spp = readRDS(here::here("outputs_figures/5_traitglms/ocean_models_PA/ocean_spp.rds"))
ocean_clusters_many = readRDS(here::here("outputs_figures/5_traitglms/ocean_models_PA/ocean_clusters_many.rds"))
ocean_trait_many = readRDS(here::here("outputs_figures/5_traitglms/ocean_models_PA/ocean_traits_many.rds"))
ocean_full_spp_many = readRDS(here::here("outputs_figures/5_traitglms/ocean_models_PA/ocean_full_spp_many.rds"))

#Anovas
an_ocean_spp = readRDS(here::here("outputs_figures/5_traitglms/ocean_models_PA/ocean_spp_anova.rds"))
an_ocean_clusters_many = readRDS(here::here("outputs_figures/5_traitglms/ocean_models_PA/ocean_clusters_many_anova.rds"))
an_ocean_trait_many = readRDS(here::here("outputs_figures/5_traitglms/ocean_models_PA/ocean_traits_many_anova.rds"))
an_ocean_full_spp = readRDS(here::here("outputs_figures/5_traitglms/ocean_models_PA/ocean_full_spp_many_anova.rds"))
```

### Species model

**SPP Manyglm - Ocean Basin**

```{r SPP ocean manyglm, eval=FALSE}

ocean_spp <- traitglm(L = alb_prey_pa_use_mv,
                      R = alb_covars_use$OceanBasin,
                      family = binomial(link = "logit")#,
                      #composition = TRUE #note that col.intercepts = TRUE is the default
                      ) #note default is "manyglm"
ocean_spp
saveRDS(ocean_spp, file = here::here("outputs_figures/5_traitglms/ocean_models_PA/ocean_spp.rds"))
```

```{r Species ocean manyglm output, eval=FALSE}
#ocean_spp$deviance
an_ocean_spp = anova(ocean_spp, test = "LR", nBoot = 100)
an_ocean_spp
#predict.traitglm(ocean_spp)
saveRDS(an_ocean_spp, file = here::here("outputs_figures/5_traitglms/ocean_models_PA/ocean_spp_anova.rds"))

#ocean_ocean_spp_sum = summary(ocean_spp, 
#                              nBoot=100,
#                              show.time=TRUE)
#ocean_ocean_spp_sum
#saveRDS(ocean_ocean_spp_sum, file = here::here("outputs_figures/5_traitglms/ocean_models_PA/ocean_spp_summary.rds"))
```

```{r SPP ocean manyglm eval, tidy=TRUE, message=FALSE, warning=FALSE}
ocean_spp
an_ocean_spp

#Check model output - residuals TEST
qqnorm(residuals(ocean_spp)[which(residuals(ocean_spp)<10000)]); abline(c(0,1),col="red")

#omit non-finite residuals FULL MODEL
resids.full = residuals(ocean_spp)[is.finite((residuals(ocean_spp)))]
plot(resids.full); abline(c(0,0),col="red")

```

### Trait guild model

**MANYGLM Clusters - Ocean Basin**

Trait (cluster)*Env (ocean basin) - Manyglm

```{r Clusters ocean manyglm, eval=FALSE}

ocean_clusters_many <- traitglm(L = alb_prey_pa_use_mv,
                                R = alb_covars_use$OceanBasin,
                                Q = traits_use$cluster,
                                family = binomial(link = "logit"),
                                composition = TRUE
                                ) 
#note default is "manyglm"
#note that col.intercepts = TRUE is the default
ocean_clusters_many
saveRDS(ocean_clusters_many, file = here::here("outputs_figures/5_traitglms/ocean_models_PA/ocean_clusters_many.rds"))
```

```{r Clusters ocean manyglm output, eval=FALSE}

#ocean_clusters_many$deviance
an_ocean_clusters_many = anova(ocean_clusters_many, test = "LR", nBoot = 100)
an_ocean_clusters_many 
saveRDS(an_ocean_clusters_many, file = here::here("outputs_figures/5_traitglms/ocean_models_PA/ocean_clusters_many_anova.rds"))

#ocean_ocean_clusters_sum = summary(ocean_clusters_many, nBoot=10)
#ocean_ocean_clusters_sum
#saveRDS(ocean_ocean_clusters_sum, file = here::here("outputs_figures/5_traitglms/ocean_models_PA/ocean_clusters_many_summary.rds"))
```

```{r Plot clusters ocean manyglm, tidy=TRUE, message=FALSE, warning=FALSE}
ocean_clusters_many
an_ocean_clusters_many

a        = max( abs(ocean_clusters_many$fourth.corner) )
colort   = colorRampPalette(c("blue","grey","red")) #can replace "grey" with "white"
temp = t(as.matrix(ocean_clusters_many$fourth.corner))

rownames(temp) = c(#"Mediterranean",
                   "N Atlantic",
                   "N Pacific"#, "S Pacific"
                   )

colnames(temp)

ocean_clusters_many_graph = levelplot(temp, 
                     xlab=list("", cex=1.5), #Ocean Basins
                     ylab=list("Trait Guilds", cex=1.5), 
                     col.regions=colort(100), 
                     at=seq(-a, a,length=100),
                     scales = list(x=list(rot = 45, cex=1.3), y=list(cex=1.3)), 
                     colorkey=list(labels=list(cex=1.3)))

print(ocean_clusters_many_graph)

ggsave(here::here("outputs_figures/5_traitglms/ocean_models_PA/ocean_clusters_many_graph.jpeg"), 
       plot = as.ggplot(ocean_clusters_many_graph), width=12, height=12, dpi=300)
ggsave(here::here("outputs_figures/5_traitglms/ocean_models_PA/ocean_clusters_many_graph.pdf"), 
       plot = as.ggplot(ocean_clusters_many_graph), width=12, height=12, dpi=300)
```

### Trait models

**MANYGLM Trait values - Ocean Basin**

Trait (trait matrix)*Env (ocean basin) - Manyglm

```{r Traits ocean manyglm, eval=FALSE}
ocean_trait_many <- traitglm(L = alb_prey_pa_use_mv,
                             R = alb_covars_use$OceanBasin,
                             Q = traits_use[,1:4],
                             family=binomial(link = "logit"),
                             composition = TRUE
                             ) 
#note default is "manyglm"
#note that col.intercepts = TRUE is the default
ocean_trait_many
saveRDS(ocean_trait_many, file = here::here("outputs_figures/5_traitglms/ocean_models_PA/ocean_traits_many.rds"))
```

```{r Traits ocean manyglm output, eval=FALSE}
#ocean_trait_many$deviance
an_ocean_trait_many = anova(ocean_trait_many, test = "LR", nBoot = 100)
an_ocean_trait_many
saveRDS(an_ocean_trait_many, file = here::here("outputs_figures/5_traitglms/ocean_models_PA/ocean_traits_many_anova.rds"))

#ocean_ocean_trait_sum = summary(ocean_trait_many, nBoot=10)
#ocean_ocean_trait_sum
#saveRDS(ocean_ocean_trait_sum, file = here::here("outputs_figures/5_traitglms/ocean_models_PA/ocean_trait_many_summary.rds"))
```

```{r Plot traits ocean manyglm, tidy=TRUE, message=FALSE, warning=FALSE}
ocean_trait_many
an_ocean_trait_many

a        = max( abs(ocean_trait_many$fourth.corner) )
colort   = colorRampPalette(c("blue","grey","red")) #can replace "grey" with "white"
temp = t(as.matrix(ocean_trait_many$fourth.corner))

rownames(temp) = c(#"Mediterranean",
                   "N Atlantic",
                   "N Pacific"#,
                   #"S Pacific"
                   )

colnames(temp) 

ocean_trait_many_graph = levelplot(temp, 
                     xlab=list("", cex=1.5), #Ocean Basins
                     ylab=list("Individual traits", cex=1.5), 
                     col.regions=colort(100), 
                     at=seq(-a, a,length=100),
                     scales = list(x=list(rot = 45, cex=1.3), y=list(cex=1.3)), 
                     colorkey=list(labels=list(cex=1.3)))


print(ocean_trait_many_graph)

ggsave(here::here("outputs_figures/5_traitglms/ocean_models_PA/ocean_traits_many_graph.pdf"), 
       plot = as.ggplot(ocean_trait_many_graph), width=12, height=12, dpi=300)
ggsave(here::here("outputs_figures/5_traitglms/ocean_models_PA/ocean_traits_many_graph.jpeg"), 
       plot = as.ggplot(ocean_trait_many_graph), width=12, height=12, dpi=300)
```


### Full model

**MANYGLM Speceis + Trait values - Ocean Basin**

```{r Full spp trait model ocean manyglm, eval=FALSE}

ocean_full_many <- traitglm(L = alb_prey_pa_use_mv,
                            R = alb_covars_use$OceanBasin,
                            Q = alb_trait_use[,c(1,4:7)], 
                            family=binomial(link = "logit")#,
                            #composition = TRUE
                            ) 
ocean_full_many
saveRDS(ocean_full_many, here::here("outputs_figures/5_traitglms/ocean_models_PA/ocean_full_spp_many.rds"))
```

```{r Full spp model output, eval=FALSE}
an_ocean_full_many = anova(ocean_full_many, nBoot = 100)
an_ocean_full_many
saveRDS(an_ocean_full_many, here::here("outputs_figures/5_traitglms/ocean_models_PA/ocean_full_spp_many_anova.rds"))

#ocean_full_spp_sum = summary(ocean_full_many, nBoot=10)
#ocean_full_spp_sum
```

### Model comparison

```{r all outputs}
an_ocean_spp
an_ocean_clusters_many
an_ocean_trait_many
```

## TRAIT GLMS - LASSO

### Read in models

```{r Read in LASSO models}
#Manyglm
ocean_spp_LASSO = readRDS(here::here("outputs_figures/5_traitglms/ocean_models_PA/ocean_spp_LASSO.rds"))
ocean_cluster_LASSO = readRDS(here::here("outputs_figures/5_traitglms/ocean_models_PA/ocean_clusters_LASSO.rds"))
ocean_trait_LASSO = readRDS(here::here("outputs_figures/5_traitglms/ocean_models_PA/ocean_traits_LASSO.rds"))
ocean_phylo_LASSO = readRDS(here::here("outputs_figures/5_traitglms/ocean_models_PA/ocean_phylo_LASSO.rds"))
```

### Species traitglm LASSO

**SPP LASSO - Ocean Basin**

```{r SPP ocean LASSO, eval=FALSE}

ocean_spp_LASSO <- traitglm(L = alb_prey_pa_use_mv,
                      R = alb_covars_use$OceanBasin,
                      family = binomial(link = "logit"),
                      method = "glm1path"#,
                      #composition = TRUE
                      ) 
saveRDS(ocean_spp_LASSO, here::here("outputs_figures/5_traitglms/ocean_models_PA/ocean_spp_LASSO.rds"))
#note default is "manyglm"
#note that col.intercepts = TRUE is the default
```

### Trait guild tratglm LASSO

**GLM1PATH Clusters - Ocean Basin**

Using method = glm1path == LASSO model.

```{r Clusters ocean LASSO, eval=FALSE}

ocean_cluster_LASSO <- traitglm(L = alb_prey_pa_use_mv,
                                R = alb_covars_use$OceanBasin,
                                Q = traits_use$cluster,
                                family=binomial(link = "logit"),
                                method="glm1path",
                                composition = TRUE
                                ) 
saveRDS(ocean_cluster_LASSO, here::here("outputs_figures/5_traitglms/ocean_models_PA/ocean_clusters_LASSO.rds"))
#note default is "manyglm"
#note that col.intercepts = TRUE is the default
```

```{r Plot clusters ocean LASSO, tidy=TRUE, message=FALSE, warning=FALSE}
a        = max( abs(ocean_cluster_LASSO$fourth.corner) )
colort   = colorRampPalette(c("blue","grey","red")) #can replace "grey" with "white"
temp = t(as.matrix(ocean_cluster_LASSO$fourth.corner))

rownames(temp) = c("Mediterranean",
                   "N Atlantic",
                   "N Pacific"#, 
                   #"S Pacific"
                   )

colnames(temp) = c("1",
                   "2",
                   "3",
                   "4",
                   "5",
                   "6",
                   "7")

ocean_cluster_LASSO_graph = levelplot(temp, 
                     xlab=list("", cex=1.5), #Ocean Basins
                     ylab=list("Trait Guilds", cex=1.5), 
                     col.regions=colort(100), 
                     at=seq(-a, a,length=100),
                     scales = list(x=list(rot = 45, cex=1.5), y=list(cex=1.5)), 
                     colorkey=list(labels=list(cex=1.5)))

print(ocean_cluster_LASSO_graph)

ggsave(here::here("outputs_figures/5_traitglms/ocean_models_PA/ocean_clusters_LASSO_graph.jpeg"), 
       plot = as.ggplot(ocean_cluster_LASSO_graph), width=8, height=8, dpi=300)
ggsave(here::here("outputs_figures/5_traitglms/ocean_models_PA/ocean_clusters_LASSO_graph.pdf"), 
       plot = as.ggplot(ocean_cluster_LASSO_graph), width=8, height=8, dpi=300)
```

### Trait information traitglm LASSO

*GLM1PATH Trait values - Ocean Basin**

Using method = glm1path == LASSO model.

```{r Traits ocean LASSO, eval=FALSE}

ocean_trait_LASSO <- traitglm(L = alb_prey_pa_use_mv,
                              R = alb_covars_use$OceanBasin,
                              Q = traits_use[,1:4], #traits_use[,1:5]
                              family=binomial(link = "logit"),
                              method="glm1path",
                              composition = TRUE
                              ) 
saveRDS(ocean_trait_LASSO, here::here("outputs_figures/5_traitglms/ocean_models_PA/ocean_traits_LASSO.rds"))
#note default is "manyglm"
#note that col.intercepts = TRUE is the default
```

```{r Plot traits ocean LASSO, tidy=TRUE, message=FALSE, warning=FALSE}

a        = max( abs(ocean_trait_LASSO$fourth.corner) )
colort   = colorRampPalette(c("blue","grey","red")) #can replace "grey" with "white"
temp = t(as.matrix(ocean_trait_LASSO$fourth.corner))

rownames(temp) = c("Mediterranean",
                   "N Atlantic",
                   "N Pacific"#,
                   #"S Pacific"
                   )

colnames(temp) = c("Vert. habitat - demersal",
                   "Vert. habitat - epipelagic",
                   "Vert. habitat - mesopelagic",
                   "Horz. habitat - coastal",
                   "Horz. habitat - continental shelf",
                   "Horz. habitat - oceanic",
                   "Diel migrant (yes)",
                   "Seasonal migrant (yes)") #,
                   #"Gregarious - schooling"

ocean_LASSO_graph = levelplot(temp, 
                     xlab=list("", cex=1.5), #Ocean Basins
                     ylab=list("Species traits", cex=1.5), 
                     col.regions=colort(100), 
                     at=seq(-a, a,length=100),
                     scales = list(x=list(rot = 45, cex=1.5), y=list(cex=1.5)), 
                     colorkey=list(labels=list(cex=1.5)))

print(ocean_LASSO_graph)

ggsave(here::here("outputs_figures/5_traitglms/ocean_models_PA/ocean_traits_LASSO_graph.pdf"), 
       plot = as.ggplot(ocean_LASSO_graph), width=12, height=8, dpi=300)
ggsave(here::here("outputs_figures/5_traitglms/ocean_models_PA/ocean_traits_LASSO_graph.jpeg"), 
       plot = as.ggplot(ocean_LASSO_graph), width=12, height=8, dpi=300)
```

### Phylo for Supp Info

**GLM1PATH Phylo - Ocean Basin**

Using method = glm1path == LASSO model.

```{r Phylo ocean LASSO, eval=FALSE}

ocean_phylo_LASSO <- traitglm(L = alb_prey_pa_use_mv,
                                R = alb_covars_use$OceanBasin,
                                Q = alb_trait_use$prey_order,
                                family=binomial(link = "logit"),
                                method="glm1path",
                                composition = TRUE
                                )
saveRDS(ocean_phylo_LASSO, file = here::here("outputs_figures/5_traitglms/ocean_models_PA/ocean_phylo_LASSO.rds"))
```

```{r Plot phylo ocean LASSO, tidy=TRUE, message=FALSE, warning=FALSE}

a        = max( abs(ocean_phylo_LASSO$fourth.corner) )
colort   = colorRampPalette(c("blue","grey","red")) #can replace "grey" with "white"
temp = t(as.matrix(ocean_phylo_LASSO$fourth.corner))

rownames(temp) = c("Mediterranean",
                   "N Atlantic",
                   "N Pacific"
                   )

colnames(temp) = c("Amphipoda", "Aulopiformes", "Beloniformes", "Caenogastropoda", "Clupeiformes", "Decapoda", "Decapodiformes",
                   "Euphausiacea", "Gadiformes", "Littorinimorpha", "Myctophiformes", "Myopsida", "Octopoda", "Oegopsida",
                   "Ophidiiformes", "Perciformes", "Pleuronectiformes", "Scorpaeniformes", "Sepiida", "Siphonophorae", "Stephanoberyciformes",
                   "Stomiiformes", "Syngnathiformes")

ocean_phylo_LASSO_graph = levelplot(temp, 
                     xlab=list("", cex=1.5), #Ocean Basins
                     ylab=list("Trait Guilds", cex=1.5), 
                     col.regions=colort(100), 
                     at=seq(-a, a,length=100),
                     scales = list(x=list(rot = 45, cex=1.5), y=list(cex=1.5)), 
                     colorkey=list(labels=list(cex=1.5)))

print(ocean_phylo_LASSO_graph)

ggsave(here::here("outputs_figures/5_traitglms/ocean_models_PA/ocean_phylo_LASSO_graph.jpeg"), 
       plot = as.ggplot(ocean_phylo_LASSO_graph), width=8, height=12, dpi=300)
ggsave(here::here("outputs_figures/5_traitglms/ocean_models_PA/ocean_phylo_LASSO_graph.pdf"), 
       plot = as.ggplot(ocean_phylo_LASSO_graph), width=8, height=12, dpi=300)
```
